% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smooth.r
\name{binning}
\alias{binning}
\title{Bin data}
\usage{
binning(IDE, DEP, MM = NULL, bins = 23, bwd = c("lin", "log10", "exp10",
  "logE", "expE")[3], mfun = c("mean", "median")[1])
}
\arguments{
\item{Currently}{none}
}
\value{
Currently none
}
\description{
Bin data.
}
\examples{
Currently none
}
\author{
Stefan Metzger \email{eddy4R.info@gmail.com}
}
\references{
Currently none
}
\seealso{
Currently none
}
\keyword{Currently}
\keyword{none}

