% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def.agr.ecte.dp01.R
\name{def.agr.ecte.dp01}
\alias{def.agr.ecte.dp01}
\title{Definition function: aggregation of ecte dp01 outputs}
\usage{
def.agr.ecte.dp01(inpList, MethSubAgr = FALSE)
}
\arguments{
\item{inpList}{a list of dp01 computed output statistics and dp01 quality flags and quality metrics over multiple aggregations periods that need to be combined.}

\item{MethSubAgr}{a logical to indicate if the subset aggregated periods should be included in the data to be combined.}
}
\value{
A dataframe of indices and corresponding times for aggregation periods.
}
\description{
Definition function to produce a dataframe of indices and corresponding times for aggregation periods.
}
\examples{
Currently none
}
\references{
License: GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007.
}
\seealso{
Currently none
}
\author{
Dave Durden \email{ddurden@battelleecology.org}
}
\keyword{NEON,}
\keyword{aggregation,}
\keyword{averaging}
\keyword{intervals}
