% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def.agr.vari.seSq.R
\name{def.agr.vari.seSq}
\alias{def.agr.vari.seSq}
\title{Definition function: Determining mean, external, internal and total variance, and squared standard error}
\usage{
def.agr.vari.seSq(data = data.frame(mean, vari))
}
\arguments{
\item{data}{Dataframe of type numeric containing column vectors \code{mean} and \code{vari} of equal length.}

\item{mean}{Vector of type numeric. Means of the variable of interest at finer resolution, e.g. minutely [user-defined].}

\item{vari}{Vector of type numeric. Variances of the variable of interest at finer resolution, e.g. minutely [user-defined^2].}
}
\value{
Returns a list containing mean \code{mean}, external variance \code{variExt}, internal variance \code{variIntl}, total variance \code{variTota}, and squared standard error \code{seSq} at coarser resolution, e.g. hourly when \code{data} contains 60 minutely values.
}
\description{
Function defintion. Calculates the mean, external, internal and total variance, and the squared standard error over all supplied input values. Useful e.g. to aggregate from daily to monthly resolution by using all provided values, or similar.
}
\examples{
def.vari.seSq.agr(data = data.frame(mean = rnorm(10), vari = rnorm(10)^2))
}
\references{
License: GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007
}
\seealso{
Currently none
}
\author{
Ke Xu \email{xuke2012abroad@gmail.com} \cr
Stefan Metzger \email{eddy4R.info@gmail.com}
}
\keyword{aggregate}
\keyword{deviation,}
\keyword{error,}
\keyword{standard}
