% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def.aply.conv.poly.R
\name{def.aply.conv.poly}
\alias{def.aply.conv.poly}
\title{Apply polynomial conversion}
\usage{
def.aply.conv.poly(data, coefPoly = c(0, 1))
}
\arguments{
\item{\code{data}}{Required. A vector or matrix or type numeric, containing the data to be converted}

\item{\code{coefPoly}}{Optional. A vector of type numeric, containing the polynomial 
coefficients (in increasing order) to apply to \code{data}. 
This is a numerical vector of [a0 a1 a2 a3 ...] signifying the 
coeficients of the equation a0 + a1*x + a2*x^2 + a3*x^3 ... , where x is the input data.
Default is c(0,1).}
}
\value{
A vector or matrix, matching the format of \code{data} containing the data with 
polynomial conversion applied.
}
\description{
Function definition. Apply a polynomial equation to data given the polynomial coefficients.
}
\examples{
Currently none
}
\author{
Cove Sturtevant \email{eddy4R.info@gmail.com}
}
\references{
license: Terms of use of the NEON FIU algorithm repository dated 2015-01-16
}
\seealso{
Currently none
}
\keyword{conversion}
\keyword{equation,}
\keyword{polynomial}

