% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def.base.ec.R
\name{def.base.ec}
\alias{def.base.ec}
\title{Base state for eddy-covariance calculation}
\usage{
def.base.ec(pos, var, AlgBase)
}
\arguments{
\item{\code{pos}}{time index used for interpolation [-]}

\item{\code{var}}{variable of interest []}

\item{\code{AlgBase}}{c("mean", "trnd", "ord03") algorithm used to determine base state, where \cr
"mean" is the simple algorithmic mean, \cr
"trnd" is the least squares linear (1st order) trend, and \cr
"ord03" is the least squares 3rd order polynomial fit}
}
\value{
Vector that contains the chosen base-signal for \code{var}.
}
\description{
Function definition. Determine base state for eddy-covariance calculation.
}
\examples{
Currently none
}
\author{
Stefan Metzger \email{eddy4R.info@gmail.com} \cr
Cove Sturtevant \email{eddy4R.info@gmail.com}
}
\references{
license: Terms of use of the NEON FIU algorithm repository dated 2015-01-16
}
\seealso{
Currently none
}
\keyword{base}
\keyword{mean}
\keyword{state,}

