% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def.bin.R
\name{def.bin}
\alias{def.bin}
\title{Definition function: Binning data}
\usage{
def.bin(
  idep,
  depe,
  RngMinMax = NULL,
  NumBin,
  WidtBin = c("lin", "log10", "exp10", "logExp", "expLog"),
  Func = c("mean", "median", "sd")
)
}
\arguments{
\item{idep}{Either a vector or matrix of class numeric or integer containing the independent variable and of the same length as \code{depe}. []}

\item{depe}{Either a vector or matrix of class numeric or integer containing the dependent variable and of the same length as \code{idep}. []}

\item{RngMinMax}{An object of class numeric or integer containing the minimum and maximum values of the independent variable. Defaults to NULL. []}

\item{NumBin}{An object of class numeric or integer containing the number of bins. []}

\item{WidtBin}{An object of class string containing the functions ("lin", "log10", "exp10", "logExp", "expLog") to determine bin width distribution of the independent variable. []}

\item{Func}{An object of class string containing the arithmetic "mean", "median", or "sd". []}
}
\value{
\code{idep} A list object of class "numeric" containing the resulted binning of independent variable and of the same length as {WidtBin} and \code{depe} a matrix containing the the resulted binning of dependent variable and of the same length as {WidtBin}. \cr
}
\description{
Function definition. Smooth data using Binning method.
}
\examples{
def.bin(idep = rnorm(5000), depe = rnorm(5000), RngMinMax = NULL, NumBin = 23, WidtBin = "log10", Func = "mean" )
def.bin(idep = rnorm(500), depe = rnorm(500), RngMinMax = c(0.1, 0.4), NumBin = 12, WidtBin = "lin", Func = "median" )
}
\references{
License: GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007
}
\seealso{
Currently none
}
\author{
Stefan Metzger \email{eddy4R.info@gmail.com} \cr
Natchaya Pingintha-Durden \email{ndurden@battelleecology.org}
}
\keyword{bin}
