% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def.conv.body.met.R
\name{def.conv.body.met}
\alias{def.conv.body.met}
\title{Coordinate transformation from CSAT3 body coordinate system to meteorological coordinate system}
\usage{
def.conv.body.met(AzSoniInst, AzSoniOfst = eddy4R.base::def.conv.unit(data =
  90, unitFrom = "deg", unitTo = "rad")$data[[1]], veloBody)
}
\arguments{
\item{\code{AzSoniInst}}{Parameter of class numeric. Azimuth direction against true north in which sonic anemometer installation (transducer array) is pointing [rad]}

\item{\code{AzSoniOfst}}{Parameter of class integer or numeric.  Azimuth Offset of meteorological x-axis against true north. That is, angle by which sonic data has to be clockwise azimuth-rotated when sonic anemometer body x-axis points perfectly north (azSonic = 0) [rad]}

\item{\code{veloBody}}{Variable of class numeric. Data frame containing wind speeds along x-axis (xaxs), y-axis (yaxs),and z-axis (zaxs) in sonic anemometer body coordinate system. For example: \code{veloBody <- data.frame(xaxs=rnorm(20), yaxs=rnorm(20), zaxs=rnorm(20))} [m s-1]}
}
\value{
Wind speed in meteorological coordinate system [m s-1]
}
\description{
Function defintion. Coordinate transformation from CSAT3 body coordinate system to meteorological coordiante system.
}
\examples{
veloIn01 <- data.frame(xaxs=rnorm(20), yaxs=rnorm(20), zaxs=rnorm(20)) 
def.conv.body.met(AzSoniInst=60, veloBody=veloIn01)
}
\author{
Stefan Metzger \email{eddy4R.info@gmail.com} \cr

Hongyan Luo \email{eddy4R.info@gmail.com}
}
\references{
License: Terms of use of the NEON FIU algorithm repository dated 2015-01-16.
}
\seealso{
Currently none
}
\keyword{anemometer,}
\keyword{coordinate,}
\keyword{sonic}
\keyword{transformation}

