% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def.dens.mole.air.R
\name{def.dens.mole.air}
\alias{def.dens.mole.air}
\title{Calculation of the molar density of the mixture of dry air and water vapor}
\usage{
def.dens.mole.air(presSum, tempMean)
}
\arguments{
\item{\code{presSum}}{A vector containing the total pressure of the air mixture, of class "numeric". [Pa]}

\item{\code{tempMean}}{A vector containing the mean temperatrue of the air mixture, of class "numeric". [K]}
}
\value{
The returned object is the the molar density of the mixture of dry air and water vapor
}
\description{
Function definition. Calculation of the molar density of the misxture of dry air and water vapor
}
\examples{
Example 1, this will cause an error message due to tempIn and tempOut have no units: 
def.dens.mole.air(presSum = 86000, tempMean = 286)
Example 2, assign values and units to variables first, the function should run ok.
presSum = 86000
tempMean = 286
attributes(presSum)$unit = "Pa"
attributes(tempMean)$unit = "K"
def.dens.mole.air(presSum, tempMean)
}
\author{
Stefan Metzger \email{eddy4R.info@gmail.com}
Hongyan Luo \email{eddy4R.info@gmail.com}
}
\references{
Currently none.
}
\seealso{
Currently none.
}
\keyword{density,}
\keyword{mole}
\keyword{pressure}
\keyword{temperature,}

