% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def.dens.pres.pois.R
\name{def.dens.pres.pois}
\alias{def.dens.pres.pois}
\title{Poisson's equation (adiabatic change) - density as function of pressure change}
\usage{
def.dens.pres.pois(dens01, pres01, pres02, Kppa = eddy4R.base::Natu$KppaDry)
}
\arguments{
\item{\code{dens01}}{Measured density, Amount per volume [same unit as returned density, e.g. kg/m3 or kmol/m3].}

\item{\code{pres01}}{Measured air pressure [same unit as reference pressure]}

\item{\code{pres02}}{Reference pressure [same unit as measured air pressure]}

\item{\code{Kppa}}{Ratio of specific gas constant to specific heat at constant pressure. Default as KppaDry [-]}
}
\value{
Densities at reference pressure [same unit as measured density]
}
\description{
Poisson's equation (adiabatic change) - density as function of pressure change.
}
\examples{
dens02 <- def.dens.pres.pois(dens01 = 1.056, pres01 = 845, pres02 = 1000, Kppa = eddy4R.base::Natu$KppaDry)
}
\author{
Stefan Metzger \email{eddy4R.info@gmail.com} \cr
Natchaya P-Durden
}
\references{
Currently none
}
\seealso{
Currently none
}
\keyword{Currently}
\keyword{none}

