% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def.dens.temp.pois.R
\name{def.dens.temp.pois}
\alias{def.dens.temp.pois}
\title{Definition function: Poisson's equation (adiabatic change) - density as function of temperature change}
\usage{
def.dens.temp.pois(
  dens01,
  temp01,
  temp02,
  Kppa = eddy4R.base::IntlNatu$KppaDry
)
}
\arguments{
\item{dens01}{Measured density, Amount per volume [same unit as returned density, e.g. kg/m3 or kmol/m3].}

\item{temp01}{Measured air temperature [K]}

\item{temp02}{Reference temperature [K]}

\item{Kppa}{Ratio of specific gas constant to specific heat at constant pressure. Default as KppaDry [-]}
}
\value{
Densities at reference temperature [same unit as measured density]
}
\description{
Poisson's equation (adiabatic change) - density as function of temperature change.
}
\examples{
dens02 <- def.dens.temp.pois(dens01 = 1.056, temp01 = 298.15, temp02 = 288.15, Kppa = eddy4R.base::IntlNatu$KppaDry)
}
\references{
License: GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007
}
\seealso{
Currently none
}
\author{
Stefan Metzger \email{eddy4R.info@gmail.com} \cr
Natchaya Pingintha-Durden \email{ndurden@battelleecology.org}
}
\keyword{Currently}
\keyword{none}
