% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def.dir.wind.R
\name{def.dir.wind}
\alias{def.dir.wind}
\title{Definition function: Wind direction mean and variance calculation using vector averaging}
\usage{
def.dir.wind(inp, MethVari = c("02StepRad", "DistAngMin", "Yama")[3])
}
\arguments{
\item{inp}{numeric vector containing instantaneous wind directions}

\item{MethVari}{This is a character string to determine the method to calculate the wind direction variance. The methods include "02StepRad", "DistAngMin", and "Yama"}
}
\value{
Mean and variance of the wind direction in meteorological coordinate system [rad]
}
\description{
Function defintion. Calculate the mean and variance of the wind direction.
}
\examples{
#Instantaneous wind direction
windDirInst <- rnorm(3600, 180, 10)
def.dir.wind(windDirInst, MethVari = "DistAngMin")
}
\references{
License: GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007. 
Yamartino, R. J. (1984) A Comparison of Several “Single-Pass” Estimators of the Standard Deviation of Wind Direction. Journal of Applied Meteorology and Climatology, 23, 1362-1366.
}
\seealso{
Currently none
}
\author{
David Durden \email{ddurden@battelleecology.org}
}
\keyword{anemometer,}
\keyword{averaging}
\keyword{direction,}
\keyword{sonic}
\keyword{vector}
\keyword{wind}
