% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def.env.glob.R
\name{def.env.glob}
\alias{def.env.glob}
\title{Definition function: Global R-environment settings for use with the eddy4R family of R-packages}
\usage{
def.env.glob()
}
\arguments{
\item{Currently}{none}
}
\value{
Sys.setlocale('LC_ALL','C') \cr
if(.Platform$OS.type == "windows") windowsFonts(times=windowsFont("TT Times New Roman"))
if(Sys.info()["sysname"] == "Darwin") quartzFonts(times = quartzFont(rep("Times-Roman", 4)))
}
\description{
Function defintion. To avoid user-location-specific dependencies, the system locale is set to the C language standard. For R in Windows and Mac the 'times' font is defined. Under Unix OS, it is not neccesary to assign font "times".
}
\examples{
Currently none
}
\author{
Stefan Metzger \email{eddy4R.info@gmail.com}
}
\references{
License: GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007. \cr
\url{http://www.inside-r.org/r-doc/base/Sys.setlocale}
}
\seealso{
Currently none
}
\keyword{environment}
\keyword{global}

