% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def.extr.var.with.unit.R
\name{def.extr.var.with.unit}
\alias{def.extr.var.with.unit}
\title{Extract variable with unit attribute from data frame}
\usage{
def.extr.var.with.unit(data, nameVar)
}
\arguments{
\item{\code{data}}{Required. A data frame with units of each variables attached as a named attribute "unit"
to the data frame.}

\item{\code{nameVar}}{Required. A string containing the name of the variable within \code{data} to extract,
along with its units}
}
\value{
The desired variable with units attached as attribute.
}
\description{
Function definition. Extracts a variable from a data frame (with units attached as attributes
at the data frame level), and attaches the variable's unit directly to the extracted variable.
}
\examples{
Currently none
}
\author{
Cove Sturtevant \email{csturtevant@neoninc.org}
}
\references{
license: Terms of use of the NEON FIU algorithm repository dated 2015-01-16
}
\seealso{
Currently none
}
\keyword{Fulcrum}

