% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def.get.site.info.R
\name{def.get.site.info}
\alias{def.get.site.info}
\title{Load pre-defined site specific information into the global environment}
\usage{
def.get.site.info(loc = c("IM", "LOS", "NR", "NS", "PF", "SERC")[6])
}
\arguments{
\item{\code{loc}}{Parameter of class character. A parameter to choose the site for which site specific information is pulled.}
}
\value{
Site specific information provided as a list \code{SiteInfo}
}
\description{
Load pre-defined site specific information into the global environment.
}
\examples{
def.get.site.info("SERC")
}
\author{
Stefan Metzger \email{eddy4R.info@gmail.com}\cr
Dave Durden
}
\references{
License: Terms of use of the NEON FIU algorithm repository dated 2015-01-16.
}
\seealso{
Currently none
}
\keyword{metadata,}
\keyword{preprocessing}
\keyword{site,}

