% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def.hdf5.copy.para.R
\name{def.hdf5.copy.para}
\alias{def.hdf5.copy.para}
\title{Definition function: Copy metadata from HDF5 input file to another HDF5 output file}
\usage{
def.hdf5.copy.para(FileInp, FileOut)
}
\arguments{
\item{FileInp}{is the input file where the parameters are being read from.}

\item{FileOut}{is the output file where the parameters are being written to.}
}
\value{
A NEON formatted HDF5 file that has parameters from the input file written to the output HDF5 file.
}
\description{
Definition function. Function recreates the metadata from an HDF5 file to another HDF5 file with the same group heirarchy structure.
}
\examples{
Currently none
}
\references{
License: GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007. \cr
NEON Algorithm Theoretical Basis Document:Eddy Covariance Turbulent Exchange Subsystem Level 0 to Level 0' data product conversions and calculations (NEON.DOC.000823) \cr
NEON Algorithm Theoretical Basis Document:Eddy Covariance Turbulent Exchange Subsystem Level 1 data product calculations (NEON.DOC.000807)
}
\seealso{
Currently none
}
\author{
David Durden \email{ddurden@battelleecology.org}
}
\keyword{ECTE}
\keyword{HDF5,}
\keyword{NEON,}
\keyword{eddy-covariance,}
