% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def.hdf5.crte.R
\name{def.hdf5.crte}
\alias{def.hdf5.crte}
\title{Definition function: Create the ECTE HDF5 file structure}
\usage{
def.hdf5.crte(Date, Site = "SERC", LvlTowr, DirOut, FileOutBase = NULL,
  MethExpd = TRUE, MethDp04 = FALSE, FileNameReadMe = NULL,
  FileNameObjDesc = NULL, LvlGasRefe = c("702_000", "704_000",
  "705_000", "706_000"))
}
\arguments{
\item{Date}{is the date for the output file being generated.}

\item{Site}{is the site for which the output file is being generated.}

\item{LvlTowr}{is the measurement level of the tower top to determine the VER number of the NEON DP naming convention.}

\item{DirOut}{is the output directory where the file being generated is stored.}

\item{FileOutBase}{character string indicating the base output filename, to which the date and package information will be appended.}

\item{MethExpd}{logical indicating if the output should be expanded or basic.}

\item{MethDp04}{logical indicating if ECTE dp04 HDF5 folder structure should be included.}

\item{FileNameReadMe}{character indicating the filename incl. absolute path to the ReadMe file for inclusion in the output HDF5 file. Defaults to \code{NULL}, which downloads the readme file from a web location.}

\item{FileNameObjDesc}{character indicating the filename incl. absolute path to the object description file for inclusion in the output HDF5 file. Defaults to \code{NULL}, which downloads the object description file from a web location.}

\item{LvlGasRefe}{character string indicating the NEON measurement horizontal and vertical location (HORVER) number as described in the NEON DP naming convention.}
}
\value{
A NEON formatted HDF5 file that is output to /code{DirOut} with a readme and object description included.
}
\description{
Definition function. Function creates the standard NEON HDF5 file structure for the ECTE data products.
}
\examples{
#Example run to create L0 gold files for 4/24/2016 at SERC
#Setting Site
Site <- "SERC"
LvlTowr <- "000_060"
FileOutBase <- "NEON.D02.SERC.DP4.00200.001.ec-flux"
MethExpd <- TRUE
#Setting Date to be processed
Date <- "2016-04-24"
#Set directory for example output to working directory
DirOut <- getwd()
#Running example
def.hdf5.crte(Date = Date, Site = Site, LvlTowr = LvlTowr, DirOut = DirOut, FileOutBase = FileOutBase, MethExpd = MethExpd)
}
\references{
License: GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007. \cr
NEON Algorithm Theoretical Basis Document:Eddy Covariance Turbulent Exchange Subsystem Level 0 to Level 0’ data product conversions and calculations (NEON.DOC.000807)
}
\seealso{
Currently none
}
\author{
David Durden \email{ddurden@battelleecology.org} \cr
Stefan Metzger \email{eddy4R.info@gmail.com}
}
\keyword{ECTE}
\keyword{HDF5,}
\keyword{NEON,}
\keyword{eddy-covariance,}
