% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def.hdf5.dp.pack.ecse.R
\name{def.hdf5.dp.pack.ecse}
\alias{def.hdf5.dp.pack.ecse}
\title{Definition function: to package NEON ECSE data product outputs to be written to HDF5 files}
\usage{
def.hdf5.dp.pack.ecse(inpList, Dp = c("Dp01", "Dp02", "Dp03", "Dp04")[1])
}
\arguments{
\item{inpList}{a list of NEON ECSE computed output statistics or quality flags and quality metrics over multiple aggregations periods that need to be combined and formatted for output to HDF5.}

\item{Dp}{which level of data product is being packaged to be written to the HDF5 file, Dp = c("Dp01", "Dp02", "Dp03", "Dp04"). Defaults to "Dp01". [-]}
}
\value{
A list of dataframes of for aggregation periods.
}
\description{
Definition function to produce a list of dataframes corresponding times for aggregation periods for NEON ECSE data product, and are packaged to be written to the HDF5 file.
}
\examples{
Currently none
}
\references{
License: GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007.
}
\seealso{
Currently none
}
\author{
Dave Durden \email{ddurden@battelleecology.org} \cr
}
\keyword{ECSE,}
\keyword{HDF5}
\keyword{NEON,}
\keyword{averaging}
\keyword{data}
\keyword{intervals,}
\keyword{products,}
