% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def.hdf5.dp.pack.R
\name{def.hdf5.dp.pack}
\alias{def.hdf5.dp.pack}
\title{Definition function: to package NEON eddy-covariance data product outputs to be written to HDF5 files}
\usage{
def.hdf5.dp.pack(inpList, MethMeas = c("ecte", "ecse")[1], time, Dp)
}
\arguments{
\item{inpList}{A list of NEON eddy-covariance computed output statistics or quality flags and quality metrics over multiple aggregations periods that need to be combined and formatted for output to HDF5.}

\item{MethMeas}{A vector of class "character" containing the name of measurement method (eddy-covariance turbulent exchange or storage exchange), MethMeas = c("ecte", "ecse"). Defaults to "ecte".}

\item{time}{Optional. If MethMeas = "ecte", a dataframe including the timeBgn and timeEnd for the aggregated periods should be included in the data to be combined.}

\item{Dp}{If MethMeas = "ecte", Dp is which data product is being packaged to be written to the HDF5 file. \cr
If MethMeas = "ecse", Dp is which level of data product is being packaged to be written to the HDF5 file, Dp = c("Dp01", "Dp02", "Dp03", "Dp04").}
}
\value{
A list of dataframes of for aggregation periods.
}
\description{
Definition function to produce a list of dataframes corresponding times for aggregation periods for NEON eddy-covariance data product, and are packaged to be written to the HDF5 file.
}
\examples{
Currently none
}
\references{
License: GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007.
}
\seealso{
Currently none
}
\author{
Natchaya Pingintha-Durden \email{ndurden@battelleecology.org} \cr
Dave Durden \email{ddurden@battelleecology.org} \cr
}
\keyword{ECSE,}
\keyword{HDF5}
\keyword{NEON,}
\keyword{averaging}
\keyword{data}
\keyword{intervals,}
\keyword{products,}
