% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def.hdf5.wrte.dp01.api.R
\name{def.hdf5.wrte.dp01.api}
\alias{def.hdf5.wrte.dp01.api}
\title{Definition function: Gather/Write reingested NEON Level 1 data, qfqm, and uncertainty to output HDF5}
\usage{
def.hdf5.wrte.dp01.api(
  date,
  FileOut,
  SiteLoca,
  DpName,
  LvlTowr,
  TimeAgr,
  Tokn = NULL
)
}
\arguments{
\item{date}{Character: The date for the data to be gathered.}

\item{FileOut}{Character: The file name for the output HDF5 file}

\item{SiteLoca}{Character: Site location.}

\item{DpName}{Character: The data product name for the data to be gathered.}

\item{LvlTowr}{Character: The tower level that the sensor data is being collected in NEON data product convention (HOR_VER).}

\item{TimeAgr}{Integer: The time aggregation index in minutes (i.e. 30)}

\item{Tokn}{Optional. Character string. An API token allowing additional permissions/access.}
}
\value{
An updated dp0p HDF5 file with dp01 data, qfqm, and uncertainty written
}
\description{
Definition function. To write NEON Level 1 data product descriptive statistics (mean, minimum, maximum, variance, number of non-NA points), quality flags and quality metrics, and uncertainty values gathered from via the API to an output HDF5 file.
}
\examples{
Currently none.
}
\references{
License: GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007.
}
\seealso{
Currently none.
}
\author{
David Durden \email{ddurden@battelleecology.org}
}
\keyword{HDF5}
\keyword{output,}
