% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def.hdf5.wrte.dp01.R
\name{def.hdf5.wrte.dp01}
\alias{def.hdf5.wrte.dp01}
\title{Definition function: Write NEON Level 1 data, qfqm, and uncertainty to output HDF5}
\usage{
def.hdf5.wrte.dp01(inpList, FileOut, SiteLoca, LevlTowr, Dp01)
}
\arguments{
\item{inpList}{A list of including dp01 data, quality flags and quality metrics, and uncertainty calculations to package and write to an output HDF5 file}

\item{FileOut}{Character: The file name for the output HDF5 file}

\item{SiteLoca}{Character: Site location.}

\item{LevlTowr}{Character: The tower level that the sensor data is being collected in NEON data product convention (HOR_VER)}
}
\value{
An HDF5 file with dp01 data, qfqm, and uncertainty written
}
\description{
Definition function. To write NEON Level 1 data product descriptive statistics (mean, minimum, maximum, variance, number of non-NA points), quality flags and quality metrics, and uncertainty quantification to an output HDF5 file.
}
\references{
License: GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007.
}
\seealso{
Currently none.
}
\author{
David Durden \email{ddurden@battelleecology.org}
}
\keyword{HDF5}
\keyword{output,}
