% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def.idx.agr.R
\name{def.idx.agr}
\alias{def.idx.agr}
\title{Definition function: indices for aggregation periods}
\usage{
def.idx.agr(time, PrdAgr, FreqLoca)
}
\arguments{
\item{time}{a vector of timestamps}

\item{PrdAgr}{the time period to aggregate to averaging in seconds (30 min = 1800 s)}

\item{FreqLoca}{the frequency of the measurements that are being aggregated in hertz (Hz)}
}
\value{
A dataframe of indices and corresponding times for aggregation periods.
}
\description{
Definition function to produce a dataframe of indices and corresponding times for aggregation periods.
}
\examples{
FreqLoca <- 20
timeMeas <- base::as.POSIXlt(seq.POSIXt(
  from = base::as.POSIXlt("2016-01-01 00:00:00.001", format="\%Y-\%m-\%d \%H:\%M:\%OS", tz="UTC"),
  to = base::as.POSIXlt("2016-01-01 05:00:00.002", format="\%Y-\%m-\%d \%H:\%M:\%OS", tz="UTC"),
  by = 1/FreqLoca), tz = "UTC")
PrdAgr <- 1800
def.idx.agr(time = timeMeas, PrdAgr = PrdAgr, FreqLoca = FreqLoca)
}
\references{
License: GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007.
}
\seealso{
Currently none
}
\author{
Dave Durden \email{ddurden@battelleecology.org}
}
\keyword{NEON,}
\keyword{aggregation,}
\keyword{averaging}
\keyword{intervals}
