% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def.idx.agr.diff.R
\name{def.idx.diff}
\alias{def.idx.diff}
\title{Definition function: indices for difference calculation, e.g. time rate of change of temperature}
\usage{
def.idx.diff(PrdWndwAgr, PrdIncrAgr, numDate)
}
\arguments{
\item{PrdWndwAgr}{the width of time period to aggregate to averaging in seconds (4 min = 240 s) [s]}

\item{PrdIncrAgr}{the incremental time period to calculate the difference in seconds (30 min = 1800 s) [s]}

\item{numDate}{the number of days in the dataset [-]}
}
\value{
A dataframe of indices for difference calculation.
}
\description{
Definition function to produce a dataframe of indices and corresponding times for difference calculation.
}
\examples{
PrdWndwAgr <- 4 * 60
PrdIncrAgr <- resoTimeDp02[[idxDp]] * 60
numDate <- 1
wrk$whrData <- eddy4R.base::def.idx.diff(
 PrdWndwAgr=PrdWndwAgr,
 PrdIncrAgr=PrdIncrAgr,
 numDate=numDate
)
}
\references{
License: GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007.
}
\seealso{
Currently none
}
\author{
Ke Xu \email{kxu@battelleecology.org}
}
\keyword{NEON,}
\keyword{aggregation,}
\keyword{averaging}
\keyword{difference}
\keyword{intervals,}
