% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def.inst.depe.R
\name{def.inst.depe}
\alias{def.inst.depe}
\title{Definition function: Install dependent packages}
\usage{
def.inst.depe(
  DirPack = ".",
  Depe = TRUE,
  Repo = base::getOption("repos")[1],
  Lib = .libPaths()[1]
)
}
\arguments{
\item{DirPack}{Path to the package \code{DESCRIPTION} file, defaults to the current working directory. Character vector with single entry [-].}

\item{Depe}{Defines what type of dependencies are installed. Defaults to \code{TRUE} which includes all of \code{c("Depends", "Imports", "LinkingTo")}. \code{c("Depends", "Imports", "LinkingTo")} can also be specified individually. Logical vector with single entry or character vector [-].}

\item{Repo}{The base URL(s) of the repositories to use, e.g., the URL of a CRAN mirror such as "https://cloud.r-project.org". Defaults to \code{base::getOption("repos")[1]}. Character vector [-].}

\item{Lib}{Path to the libraries in which the dependent packages are installed. Character vector [-].}
}
\description{
Function defintion. Installs dependent packages from a source package \code{DESCRIPTION} file, without installing the source package itself. Useful e.g. for re-packing source packages with Roxygen, because devtools::document() requires all dependent packages to be installed. In case during development new dependent packages have been added to the \code{DESCRIPTION} file but are not yet installed, devtools::document() stops with an error. Calling eddy4R.base::def.inst.depe() prior to devtools::document() mitigates this error.
}
\examples{
eddy4R.base::def.inst.depe()
}
\references{
License: GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007. \cr
\url{https://gist.github.com/jtilly/ac1b028f3666ce1d82c2}. \cr
}
\seealso{
Currently none.
}
\author{
Stefan Metzger \email{eddy4R.info@gmail.com}
}
\keyword{Roxygen}
\keyword{dependencies,}
\keyword{devtools,}
\keyword{document,}
\keyword{error,}
\keyword{package,}
