% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def.irga.vali.cor.R
\name{def.irga.vali.cor}
\alias{def.irga.vali.cor}
\title{definition function: Calculation of time-series of the correction IRGA sub data products.}
\usage{
def.irga.vali.cor(
  data,
  DateProc,
  coef,
  valiData,
  valiCrit = FALSE,
  ScalMax = 20,
  FracSlpMax = 0.1,
  Freq = 20
)
}
\arguments{
\item{data}{List consisting of \code{ff::ffdf} file-backed objects containing the dp0p input IRGA.}

\item{DateProc}{A vector of class "character" containing the processing date.}

\item{coef}{List consists of linear regression coefficients (slope and offset) for DateProc - 1, DateProc, and DateProc + 1.}

\item{valiData}{List consisting of descriptive statistics (mean, min, max, vari, numSamp, se) of CO2 dry mole concentration during performing validation and CO2 dry mole concentration of reference gases.}

\item{valiCrit}{A logical stating if there are more than one validation occurred within DateProc. Defaut to FALSE.}

\item{ScalMax}{Maximum scale value (resulted from maximum-likelihood fitting of a functional relationship (MLFR)). Defaults to 20.}

\item{FracSlpMax}{Maximum fraction of slope value (resulted from maximum-likelihood fitting of a functional relationship (MLFR)). Defaults to 0.1.}

\item{Freq}{Measurement frequency. Defaults to 20. [Hz]}
}
\value{
The returned dataframe consists of the correction IRGA sub data products.
}
\description{
Definition function to calculate time-series of the correction IRGA sub data products based on linear interpolation method.
}
\examples{
Currently none.
}
\references{
License: GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007.
}
\seealso{
Currently none.
}
\author{
Natchaya P-Durden \email{eddy4R.info@gmail.com}
}
\keyword{derived,}
\keyword{irgaTurb,}
\keyword{post-processing,}
\keyword{pre-processing,}
\keyword{validation}
