% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def.log.init.R
\name{def.log.init}
\alias{def.log.init}
\title{Initialize logging for NEON IS Data Processing}
\usage{
def.log.init(Name = NULL, To = NULL, NameFile = NULL, Fmt = NULL, Lvl = NULL)
}
\arguments{
\item{Name}{String (optional). Name of logger. Defaults to root if environment variable 
LOG_NAME is not found.}

\item{To}{String (optional). Location to log to. Options are 'stdout' (console) or 'file'. 
Defaults to stdout if environment variable LOG_TO is not found.}

\item{NameFile}{String (optional). Name of file, including relative or absolute path, to log to. 
Ingored if To='stdout'. Defaults to ./log.txt if environment variable LOG_FILENAME is not found.}

\item{Fmt}{String (optional). Logging format. Options are 'plain' (human readable text)
or 'json'. Defaults to plain if environment variable LOG_FORMAT is not found.}

\item{Lvl}{String (optional). The minimum logging level that triggers log output. Options are, in 
order of increasing severity, 'debug','info','warn','error','fatal'. 
Defaults to info if environment variable LOG_LEVEL is not found.}
}
\value{
A logger object, configured according to inputs. Logging can now be conducted in the main 
code, using the functions in the lgr package. See example.
}
\description{
Definition function. Initialize logging for NEON IS Data Processing. Currently uses the lgr 
package and implements hierarchical logging with options for json formatting and writing to file.
Note that all parameters default to NULL, in which case an attempt will be made to retrieve the 
system environmental variables specified.
}
\examples{
lg <- def.log.init(To='stdout',Fmt='plain','Lvl'='info')
lg$warn('This is a warning')
lg$debug('This debug message will not print because the minimum logging level is info')
}
\references{
License: (example) GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007
}
\seealso{
Currently none
}
\author{
Cove Sturtevant \email{csturtevant@battelleecology.org}
}
\keyword{Currently}
\keyword{none}
