% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maxcor.r
\name{def.max.cor}
\alias{def.max.cor}
\title{Lag two datasets, so as to maximise their correlations}
\usage{
def.max.cor(ref, obs, refData = ref, obsData = obs, obsVars = NULL,
  maxLag = 2 * freqLoc, hardLag = FALSE, npLag = c("n", "p", "np")[3],
  absolute = FALSE, freqLoc = freq_res, hpf = TRUE, FRACmin = 0.1)
}
\arguments{
\item{Currently}{none}
}
\value{
Currently none
}
\description{
Lag two datasets, so as to maximise their correlations.
}
\examples{
Currently none
}
\author{
Stefan Metzger \email{eddy4R.info@gmail.com}\n
Dave Durden
}
\references{
Currently none
}
\seealso{
Currently none
}
\keyword{Currently}
\keyword{none}

