% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def.mean.med.mode.R
\name{def.mean.med.mode}
\alias{def.mean.med.mode}
\title{Definition function: Calculate descriptive statistics based on arithmetic mean, median and mode}
\usage{
def.mean.med.mode(test, refe = 0, Perc = FALSE)
}
\arguments{
\item{\code{test}}{A vector containing the test data. Of class "numeric" or "integer", and of the same length as \code{refe}. [same units as reference data]}

\item{\code{refe}}{A vector containing the reference data. Of class "numeric" or "integer". Defaults to \code{refe = 0}. [same units as test data]}

\item{\code{Perc}}{Report results in the same units as \code{refe} and \code{test} (\code{Perc = FALSE}) or as percentages (\code{Perc = TRUE})? Of class "logical", defaults to \code{Perc = FALSE}.}
}
\value{
\code{statLoc} A list object of class "numeric" [1, 1:3] containing mean, median, and mode. \cr

\code{statDis} A list object of class "numeric" [1, 1:3] containing standard deviation, median absolute deviation, and mode absolute deviation. \cr

\code{NumSamp} Sample size which not included NaNs. \cr
}
\description{
Calculate descriptive statistics based on arithmetic mean, median and mode.
}
\examples{
#works
def.mean.med.mode(test = rnorm(100))

#returns error message
def.mean.med.mode(test = rnorm(10), refe = 0, Perc = TRUE)
}
\references{
License: GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007 \cr
Croux, C., and Rousseeuw, P. J.: Time-efficient algorithms for two highly robust estimators of scale, Computational Statistics, 1, 411-428, 1992. \cr
http://stackoverflow.com/questions/2547402/standard-library-function-in-r-for-finding-the-mode \cr
}
\seealso{
\code{\link[eddy4R.base]{def.mode}}
}
\author{
Stefan Metzger \email{eddy4R.info@gmail.com} \cr
Natchaya Pingintha-Durden \email{ndurden@battelleecology.org}
}
\keyword{descriptive}
\keyword{statistics}
