% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def.med.mad.R
\name{def.med.mad}
\alias{def.med.mad}
\title{Definition function: Median and median absolute deviation as robust measures of scale and dispersion}
\usage{
def.med.mad(test, refe = 0, Perc = FALSE)
}
\arguments{
\item{\code{test}}{A vector containing the test data. Of class "numeric" or "integer", and of the same length as \code{refe}. [same units as reference data]}

\item{\code{refe}}{A vector containing the reference data. Of class "numeric" or "integer", and of the same length as \code{test}. Defaults to \code{refe = 0}. [same units as test data]}

\item{\code{Perc}}{Report results in the same units as \code{refe} and \code{test} (\code{Perc = FALSE}) or as percentages (\code{Perc = TRUE})? Of class "logical", defaults to \code{Perc = FALSE}.}
}
\value{
Returns object of class "numeric" [1, 1:3] containing median, median absolute deviation, and sample size (NaNs not included).
}
\description{
Function defintion. Calculates the median as robust measure of scale, and the sample-size corrected median absolute deviation as robust measure of dispersion. By default, the univariate statistics of the test data \code{test} are returned. If reference data \code{refe} is provided, the bivariate statistics of the residuals \code{test - refe} are returned.
}
\examples{
#works
def.med.mad(test = rnorm(500))
def.med.mad(test = rnorm(10), refe = rnorm(10))

#returns error message
def.med.mad(test = rnorm(10), refe = rnorm(5000))
}
\references{
License: GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007. \cr
Croux, C., and Rousseeuw, P. J.: Time-efficient algorithms for two highly robust estimators of scale, Computational Statistics, 1, 411-428, 1992. \cr
}
\seealso{
Function eddy4R.base:::def.rmsd.bias.prec.cdet() for Gaussian statistics-based measures of scale and dispersion.
}
\author{
Stefan Metzger \email{eddy4R.info@gmail.com}
}
\keyword{robust}
\keyword{statistics}
