% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def.mode.R
\name{def.mode}
\alias{def.mode}
\title{Definition function: Calculate mode based on a continuous distribution}
\usage{
def.mode(x)
}
\arguments{
\item{x}{Either a vector or an object of class numeric of the data from which the estimate is to be computed.}
}
\value{
The estimated arithmetic mode of the values in \code{x}. \cr
}
\description{
Calculate mode based on a continuous distribution.
}
\examples{
def.mode(x = rnorm(100))
}
\references{
License: GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007 \cr
\url{http://stackoverflow.com/questions/2547402/standard-library-function-in-r-for-finding-the-mode} \cr
}
\seealso{
Currently none
}
\author{
Stefan Metzger \email{eddy4R.info@gmail.com} \cr
Natchaya Pingintha-Durden \email{ndurden@battelleecology.org}
}
\keyword{descriptive}
\keyword{mode}
\keyword{statistics,}
