% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def.mtch.out.refe.R
\name{def.mtch.out.refe}
\alias{def.mtch.out.refe}
\title{Compare output against reference}
\usage{
def.mtch.out.refe(fileOut, fileRefe, Head = FALSE, NumLine = 10)
}
\arguments{
\item{\code{fileOut}}{String. The relative or absolute path and filename of the numerical data output.}

\item{\code{fileRefe}}{String. The relative or absolute path and filename of the numerical reference data.}

\item{\code{Head}}{Logical. TRUE if there is 1 header row in both the output and reference files. Defaults to FALSE.}

\item{\code{NumLine}}{The number of output rows to compare (starting from the first row). Defaults to 10.}
}
\value{
Currently none
}
\description{
Function definition. A function to compare the file output from a workflow to a reference output file
to ensure that changes during development to functions called by the workflow have not impacted the 
results.\cr
The function reads in the new data produced in the tmp directory and compares the first 10 lines to the reference output data.
}
\examples{
Currently none
}
\author{
Dave Durden \email{eddy4R.info@gmail.com} 
Cove Sturtevant \email{eddy4R.info@gmail.com}
}
\references{
License: Terms of use of the NEON FIU algorithm repository dated 2015-01-16. \cr
}
\seealso{
Currently none
}
\keyword{file}
\keyword{function}
\keyword{gold}
\keyword{test,}

