% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def.neon.dp01.R
\name{def.neon.dp01}
\alias{def.neon.dp01}
\title{Create NEON Level 1 data product descriptive statistics}
\usage{
def.neon.dp01(data = data.frame(data), vrbs = FALSE)
}
\arguments{
\item{\code{data}}{A numeric vector containing the L0' (calibrated) input data at native resolution. Of class numeric". [-]}
}
\value{
A data frame of descriptive statistics:\cr
\code{mean} The mean of non-NA values in \code{data}
\code{min} The minimum value of non-NA values in \code{data}
\code{max} The maximum value of non-NA values in \code{data}
\code{vari} The variance of non-NA values in \code{data}
\code{num} The number of non-NA values in \code{data}
\code{se} The standard error of the mean of non-NA values in \code{data}
}
\description{
Function definition. Compute NEON Level 1 data product descriptive statistics (mean, minimum, maximum, variance, number of non-NA points) by aggregating the input data over its entire range.
}
\examples{
data <- c(1,2,3,NA,5,6,7,NaN,9,10) # Calibrated raw data
dp01 <- def.dp01.stat(data=data) # Level 1 descriptive statistics
}
\author{
Cove Sturtevant \email{csturtevant@neoninc.org} \cr
}
\references{
NEON.DOC.003311 - NEON DATA PRODUCTS DEVELOPMENT PLAN
}
\seealso{
Currently none.
}
\keyword{aggregate,}
\keyword{average,}
\keyword{descriptive}
\keyword{statistics}

