% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def.para.flow.R
\name{def.para.flow}
\alias{def.para.flow}
\title{Definition function: Determine the workflow variables}
\usage{
def.para.flow(Deve = TRUE, DirFilePara = NULL, DirInp = NA, DirMnt = NA,
  DirOut = NA, DirTmp = NA, DirWrk = NA, Dom = NA, FileDp0p = NULL,
  Loc = NULL, Read = "hdf5", VersDp = c("001", "004")[1],
  VersEddy = "latest", MethParaFlow = c("DfltInp", "EnvVar")[1],
  MethMeas = c("ecte", "ecse")[1], ...)
}
\arguments{
\item{Deve}{is logical that determines if only a subset of the data should be read in to reduce testing time during development (\code{Deve = TRUE}) or all the input data should be read in (\code{Deve = FALSE})}

\item{DirFilePara}{is file path for the hdf5 dp0p input data file}

\item{DirInp}{is directory path for the hdf5 dp0p input data file}

\item{DirMnt}{is the base directory path for where the docker is mounted}

\item{DirOut}{is directory path for the output data}

\item{DirTmp}{is directory path for temporary storage during processing}

\item{DirWrk}{is directory path for working storage during processing}

\item{Dom}{is the NEON domain the site is located in}

\item{FileDp0p}{is a character string that lists the dates to be processed}

\item{Loc}{is site where the data was collected (NEON 4 letter code, e.g. SERC)}

\item{Read}{determine if the data are read from hdf5 dp0p input data file or other input files}

\item{VersDp}{is the data product level that will be output}

\item{VersEddy}{is the version of the eddy4R docker that is being used to perform the processing}

\item{MethParaFlow}{is the method used to specify workflow parameters, "EnvVar" will grab ParaFlow parameters from environmental variable and "DfltInp" will use whatever is specified in the function call.}

\item{MethMeas}{A vector of class "character" containing the name of measurement method (eddy-covariance turbulent exchange or storage exchange), MethMeas = c("ecte", "ecse"). Defaults to "ecte".}
}
\value{
\code{ParaFlow} is a list returned that indicates the workflow control parameters, including \code{ParaFlow$DirFilePara},\code{ParaFlow$DirInp}, \code{ParaFlow$DirMnt}, \code{ParaFlow$DirOut}, \code{ParaFlow$DirTmp}, \code{ParaFlow$DirWrk},\code{ParaFlow$Dom}, \code{ParaFlow$FileDp0p}, \code{ParaFlow$Loc},  \code{ParaFlow$Read}, \code{ParaFlow$VersDp}, \code{ParaFlow$VersEddy}.
}
\description{
Definition function. Function to determine the workflow variables by either reading them in from the environmental variables, defining them explicitly, or defining them by default values
}
\examples{
def.para.flow(DirFilePara = "test.h5", MethParaFlow = "DfltInp")
}
\references{
License: GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007. \cr
NEON Algorithm Theoretical Basis Document:Eddy Covariance Turbulent Exchange Subsystem Level 0 to Level 0’ data product conversions and calculations (NEON.DOC.000823) \cr
NEON Algorithm Theoretical Basis Document:Eddy Covariance Turbulent Exchange Subsystem Level 1 data product calculations (NEON.DOC.000807)
}
\seealso{
Currently none
}
\author{
David Durden \email{ddurden@battelleecology.org}
}
\keyword{ECTE}
\keyword{NEON,}
\keyword{eddy-covariance,}
\keyword{environment}
\keyword{variables,}
