% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def.para.levl.towr.R
\name{def.para.levl.towr}
\alias{def.para.levl.towr}
\title{Definition function: Determine tower top measurement level from input dp0p ECTE HDF5 file}
\usage{
def.para.levl.towr(FileIn)
}
\arguments{
\item{FileIn}{is the input dp0p ECTE HDF5 file where the parameters are being read from.}
}
\value{
\code{LevlTowr} is returned that indicates the Horizontal and Vertical indices of the tower top measurement level.
}
\description{
Definition function. Function determines tower top measurement level from input dp0p ECTE HDF5 file structure
}
\references{
License: GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007.
NEON Algorithm Theoretical Basis Document:Eddy Covariance Turbulent Exchange Subsystem Level 0 to Level 0’ data product conversions and calculations (NEON.DOC.000823)
NEON Algorithm Theoretical Basis Document:Eddy Covariance Turbulent Exchange Subsystem Level 1 data product calculations (NEON.DOC.000807)
}
\seealso{
Currently none
}
\author{
David Durden \email{ddurden@battelleecology.org}
}
\keyword{ECTE}
\keyword{HDF5,}
\keyword{NEON,}
\keyword{eddy-covariance,}
