% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def.para.thrs.R
\name{def.para.thrs}
\alias{def.para.thrs}
\title{Defination Function to read ECSE threshold table from CI-Parameter-Repo}
\usage{
def.para.thrs(dataIn, site)
}
\description{
Defination Function to read ECSE threshold table from CI-Parameter-Repo
}
\seealso{
Currently none
}
\author{
Natchaya Pingintha-Durden \email{ndurden@battelleecology.org}
}
