% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def.para.thsh.R
\name{def.para.thsh}
\alias{def.para.thsh}
\title{Definition function to read threshold table from CI-Parameter-Repo}
\usage{
def.para.thsh(dataIn, site)
}
\arguments{
\item{dataIn}{data.frame containing the threshold data from the CI-Parameter-Repo}

\item{site}{The site for which the parameters are being pulled}
}
\value{
A data.frame consisting of the threshold values to be used for the provided site.
}
\description{
Definition function to read threshold table from CI-Parameter-Repo
}
\examples{
Currently none
}
\references{
License: GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007. \cr
}
\seealso{
Currently none
}
\author{
Natchaya Pingintha-Durden \email{ndurden@battelleecology.org}
David Durden \email{ddurden@battelleecology.org}
}
\keyword{threshold}
