% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def.pois.dens.temp.R
\name{def.pois.dens.temp}
\alias{def.pois.dens.temp}
\title{Poisson's equation (adiabatic change) - density as function of temperature change}
\usage{
def.pois.dens.temp(dens01, temp01, temp02, Kppa = eddy4R.base::Natu$KppaDry)
}
\arguments{
\item{\code{dens01}}{Measured density, Amount per volume [same unit as returned density, e.g. kg/m3 or kmol/m3].}

\item{\code{temp01}}{Measured air temperature [K]}

\item{\code{temp02}}{Reference temperature [K]}

\item{\code{Kppa}}{Ratio of specific gas constant to specific heat at constant pressure. Default as KppaDry [-]}
}
\value{
Densities at reference temperature [same unit as measured density]
}
\description{
Poisson's equation (adiabatic change) - density as function of temperature change.
}
\examples{
dens02 <- def.pois.dens.temp(dens01 = 1.056, temp01 = 298.15, temp02 = 288.15, Kppa = eddy4R.base::Natu$KppaDry)
}
\author{
Stefan Metzger \email{eddy4R.info@gmail.com} \cr
Natchaya P-Durden
}
\references{
Currently none
}
\seealso{
Currently none
}
\keyword{Currently}
\keyword{none}

