% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def.pol.cart.R
\name{def.pol.cart}
\alias{def.pol.cart}
\title{Definition function: Composing azimuth angle from cartesian vector data}
\usage{
def.pol.cart(cart)
}
\arguments{
\item{\code{cart}}{variable of type numeric, matrix consisting of unit vectors X (1st column, -1 ... 1, positive to north) and Y (2nd column, -1 ... 1, positive to east)}
}
\value{
clockwise azimuth angle with 0 / 360 degree discontinuity in north [decimal degrees]
}
\description{
Composing azimuth angle from cartesian vector data.
}
\examples{
Currently none
}
\author{
Stefan Metzger \email{eddy4R.info@gmail.com} \cr
Natchaya Pingintha-Durden \email{ndurden@battelleecology.org}
}
\references{
License: GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007
}
\seealso{
Currently none
}
\keyword{unit}
\keyword{vector}

