% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def.pres.h2o.rtio.mass.h2o.dry.pres.R
\name{def.pres.h2o.rtio.mass.h2o.dry.pres}
\alias{def.pres.h2o.rtio.mass.h2o.dry.pres}
\title{Definition function: Calculate water vapor pressure from dry mass fraction and static pressure}
\usage{
def.pres.h2o.rtio.mass.h2o.dry.pres(rtioMassDryH2o, pres)
}
\arguments{
\item{rtioMassDryH2o}{Either a vector or an object of class numeric of water dry mass fraction and of the same length as \code{pres}. [kg kg-1]}

\item{pres}{Either a vector or an object of class numeric of static pressure and of the same length as \code{rtioMassDryH2o}. [Pa]}
}
\value{
Water vapor pressure and of the same length as \code{rtioMassDryH2o} and \code{pres}. [Pa]
}
\description{
Calculate water vapor pressure from dry mass fraction and static pressure.
}
\examples{
def.pres.h2o.rtio.mass.h2o.dry.pres(rtioMassDryH2o = 2.144*1e-2, pres = 65000) 
def.pres.h2o.rtio.mass.h2o.dry.pres(rtioMassDryH2o = c(2.144*1e-2, 2.617*1e-3, 4.931*1e-4), pres = c(65000, 83000, 110000)) 
}
\references{
License: GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007
}
\seealso{
Currently none
}
\author{
Stefan Metzger \email{eddy4R.info@gmail.com} \cr
Natchaya Pingintha-Durden \email{ndurden@battelleecology.org}
}
\keyword{conversion}
\keyword{humidity}
