% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def.pres.h2o.rtio.mole.h2o.dry.pres.R
\name{def.pres.h2o.rtio.mole.h2o.dry.pres}
\alias{def.pres.h2o.rtio.mole.h2o.dry.pres}
\title{Definition function: Calculate water vapor pressure from dry mole fraction and static pressure}
\usage{
def.pres.h2o.rtio.mole.h2o.dry.pres(rtioMoleDryH2o, pres)
}
\arguments{
\item{rtioMoleDryH2o}{Either a vector or an object of class numeric of water dry mole fraction and of the same length as \code{pres}. [kmol kmol-1]}

\item{pres}{Either a vector or an object of class numeric of static pressure and of the same length as \code{rtioMoleDryH2o}. [Pa]}
}
\value{
Water vapor pressure and of the same length as \code{rtioMoleDryH2o} and \code{pres}. [Pa]
}
\description{
Calculate water vapor pressure from dry mole fraction and static pressure.
}
\examples{
def.pres.h2o.rtio.mole.h2o.dry.pres(rtioMoleDryH2o = 6.52*1e-6, pres = 83000) 
def.pres.h2o.rtio.mole.h2o.dry.pres(rtioMoleDryH2o = c(6.52*1e-6, 11.617*1e-6, 4.931*1e-6), pres = c(83000, 65000, 110000)) 
}
\references{
License: GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007
}
\seealso{
Currently none
}
\author{
Stefan Metzger \email{eddy4R.info@gmail.com} \cr
Natchaya Pingintha-Durden \email{ndurden@battelleecology.org}
}
\keyword{conversion}
\keyword{humidity}
