% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def.pres.h2o.temp.R
\name{def.pres.h2o.temp}
\alias{def.pres.h2o.temp}
\title{Calculate water vapor pressure from absolute humidity and ambient temperature}
\usage{
def.pres.h2o.temp(densMassH2o, temp)
}
\arguments{
\item{\code{densMassH2o}}{Either a vector or an object of class numeric of absolute humidity and of the same length as \code{temp}. [kg m-3]}

\item{\code{temp}}{Either a vector or an object of class numeric of measured air temperature and of the same length as \code{densMassH2o}. [K]}
}
\value{
Water vapor pressure and of the same length as \code{densMassH2o} and \code{temp}. [Pa]
}
\description{
Calculate water pressure from absolute humidity and ambient temperature.
}
\examples{
def.pres.h2o.temp(densMassH2o = 1.493*1e-2, temp = 298.15)
def.pres.h2o.temp(densMassH2o = c(3.451*1e-4, 2.717*1e-3, 1.493*1e-2), temp = c(265.15, 278.15, 298.15))
}
\author{
Stefan Metzger \email{eddy4R.info@gmail.com} \cr
Natchaya Pingintha-Durden
}
\references{
License: Terms of use of the NEON FIU algorithm repository dated 2015-01-16
}
\seealso{
Currently none
}
\keyword{conversion}
\keyword{humidity}

