% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def.pres.sum.R
\name{def.pres.sum}
\alias{def.pres.sum}
\title{Calculation of total pressure in LI-7200 IRGA cell}
\usage{
def.pres.sum(presAtm, presDiff)
}
\arguments{
\item{\code{presAtm}}{A vector containing the atmospheric pressure (or static pressure), of class "numeric". [Pa]}

\item{\code{presDiff}}{A vector containing the differential pressure, of class "numeric". [Pa]}
}
\value{
The returned object is the total pressure calculated by summing the static pressure (presAtm) and the differential pressure (presDiff).
}
\description{
Function definition. Calculation of total pressure from static pressure and differential pressure
}
\examples{
Example 1, this will cause an error message due to presAtm and presDiff have no units: 
def.pres.sum(presAtm = 99380, presDiff = -1109)
Example 2, assign valuea and units to variables first, the function should run ok.
presAtm = 99380
presDiff = -1109
attributes(presAtm)$unit = "Pa"
attributes(presDiff)$unit = "Pa"
def.pres.sum(presAtm, presDiff)
}
\author{
Stefan Metzger \email{eddy4R.info@gmail.com}
Hongyan Luo \email{eddy4R.info@gmail.com}
}
\references{
Currently none.
}
\seealso{
Currently none.
}
\keyword{pressure}

