% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def.repl.char.R
\name{def.repl.char}
\alias{def.repl.char}
\title{Replace character vector elements with entries from a lookup table}
\usage{
def.repl.char(data, ReplFrom, ReplTo)
}
\arguments{
\item{\code{data}}{Named character vector, can contain values that are to be replaced as specified via \code{ReplFrom} and \code{ReplTo}.}

\item{\code{ReplFrom}}{Character vector of the same length as \code{ReplTo}, specifying the values in \code{data} that are to be replaced with the corrresponding values in \code{ReplTo}.}

\item{\code{ReplTo}}{Character vector of the same length as \code{ReplFrom}, specifying the replacement values for \code{ReplFrom} in \code{data}.}
}
\value{
Named character vector of the same length as \code{data}, with values specified in \code{ReplFrom} replaced by the corresponding values in \code{ReplTo}. If none of the entries in \code{ReplFrom} exist in \code{data}, then \code{data} is returned unchanged.
}
\description{
Function defintion. If any of the values in \code{ReplFrom} exist in \code{data}, these values are replaced with the corresponding values in \code{ReplTo}. If none of the values in \code{ReplFrom} exist in \code{data}, then \code{data} is returned unchanged.
}
\examples{
def.repl.char(
data = structure(.Data = c("Celsius", "Pa", "percent"), names = c("temperature", "pressure", "humidity")),
ReplFrom = c("Celsius", "percent"),
ReplTo = c("C", "\%")
)
}
\author{
Stefan Metzger \email{eddy4R.info@gmail.com}
}
\references{
License: Terms of use of the NEON FIU algorithm repository dated 2015-01-16. \cr
}
\seealso{
Currently none.
}
\keyword{character,}
\keyword{lookup}
\keyword{replacement,}
\keyword{table}

