% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def.repl.na.R
\name{def.repl.na}
\alias{def.repl.na}
\title{Defination Function to combine two daily files into one file and replacing NA values with previous value.}
\usage{
def.repl.na(dataProcDate, dataBgnDate, Date, numCol)
}
\arguments{
\item{\code{dataProcDate}}{A dataframe containing data and time in the processing date. [User-defined]}

\item{\code{dataBgnDate}}{A dataframe containing data and time in the day before processing date. [User-defined]}

\item{\code{Date}}{Character: Processing date e.g. "20170521". [-]}

\item{\code{numCol}}{Numeric: Define which column number in \code{dataProcDate} and \code{dataBgnDate} are being process. [-]}
}
\value{
A dataframe including the data and time in the processing date which NA values have been replaced by previous value. [User-defined]
}
\description{
Defination Function to two combine daily files into one file, replacing NA with previous value, and divide file back into daily.\cr
Specific for pump and soleniod valves.
}
\examples{
Currently none
}
\references{
License: Terms of use of the NEON FIU algorithm repository dated 2015-01-16. \cr
}
\seealso{
Currently none
}
\author{
Natchaya Pingintha-Durden \email{ndurden@battelleecology.org}
}
\keyword{NA}
\keyword{NEON,}
\keyword{eddy-covariance,}
