% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def.rh.pres.h2o.pres.sat.h2o.R
\name{def.rh.pres.h2o.pres.sat.h2o}
\alias{def.rh.pres.h2o.pres.sat.h2o}
\title{Definition function: Calculation of RH from water vapor partial pressure and saturation pressure}
\usage{
def.rh.pres.h2o.pres.sat.h2o(presH2o, presH2oSat)
}
\arguments{
\item{\code{presH2o}}{A vector containing the water vapor partial pressure, of class "numeric". [Pa]}

\item{\code{presH2oSat}}{A vector containing the water vapor saturation pressure, of class "numeric". [Pa]}
}
\value{
The returned object is the RH calculated from water vapor partial pressure and saturation pressure
}
\description{
Function definition. Calculation of RH from water vapor partial pressure and saturation pressure
}
\examples{
Example 1, this will cause an error message due to presH2o and presH2oSat have no units: 
def.rh.pres.h2o.pres.sat.h2o(presH2o = 1020, presH2oSat = 2500)
Example 2, assign values and units to variables first, the function should run ok.
presH2o = 1020
presH2oSat = 2500
attributes(presH2o)$unit = "Pa"
attributes(presH2oSat)$unit = "Pa"
def.rh.pres.h2o.pres.sat.h2o(presH2o, presH2oSat)
}
\references{
License: GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007.
}
\seealso{
Currently none.
}
\author{
Stefan Metzger \email{eddy4R.info@gmail.com} \cr
Hongyan Luo \email{hluo@battelleecology.org}
}
\keyword{RH,}
\keyword{irga}
\keyword{pressure,}
\keyword{saturation}
\keyword{vapor,partial}
\keyword{water}
