% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def.rmsd.diff.prcs.rsq.R
\name{def.rmsd.diff.prcs.rsq}
\alias{def.rmsd.diff.prcs.rsq}
\title{Definition function: RMSD, bias, precision and coefficient of determination - incl. deadband}
\usage{
def.rmsd.diff.prcs.rsq(refe, test, Perc = FALSE, Deba = NULL,
  DebaRltv = FALSE)
}
\arguments{
\item{\code{refe}}{Variable of class numeric. Reference data.Same unit as test data.}

\item{\code{test}}{Variable of class numeric. Test data.Same unit as reference data.}

\item{\code{Perc}}{Variable of class logical. It describe if the output is in percentage.}

\item{\code{Deba}}{Variable of class numeric. Numbers of the deadband around zero denominator}

\item{\code{DebaRltv}}{Variable of class logical.Absolute or relative (percentage) deadband around zero. True indicates Deba in relative (percentage), False indicates Deba in absolute numbers.}
}
\value{
Currently none
}
\description{
Function defintion.RMSD, bias, precision and coefficient of determination - incl. deadband.
}
\examples{
#refe dataset
df1 <- c(runif(200, min=-100, max=400))
#test dataset
df2 <- c(runif(200, min=-100, max=400))
#run function
def.rmsd.diff.prcs.rsq(refe=df1,test=df2,Perc = FALSE,Deba=NULL, DebaRltv=FALSE)
}
\references{
License: GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007.
}
\seealso{
Currently none
}
\author{
Stefan Metzger \email{eddy4R.info@gmail.com} \cr
Hongyan Luo \email{hluo@battelleecology.org}
}
\keyword{R-square}
\keyword{RMSD,}
\keyword{bias,}
\keyword{precision,}
