% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def.rsmp.R
\name{def.rsmp}
\alias{def.rsmp}
\title{Resampling function}
\usage{
def.rsmp(data, FreqInp, FreqOut, MethRsmp, ColDisc = NULL)
}
\arguments{
\item{\code{data}}{A vector or matrix of type numeric, containing the data to be resampled. [user-defined]}

\item{\code{FreqInp}}{Input sampling or measurements frequency. Of class "numeric" or "integer". [Hz]}

\item{\code{FreqOut}}{Desired out put frequency. Of class "numeric" or "integer". [Hz]}

\item{\code{MethRsmp}}{An object of class "character" containing the resampling method ("zoo","filt"). [-]}

\item{\code{ColDisc}}{Specific column in \code{data} which containing the discontinuity variable e.g. azimuth angle of wind direction. Of class "numeric" (column number) or "character" (column name). Defaults to NULL. [-]}
}
\value{
Data frame of resampling data and of the same number of variables as \code{data}. [user-defined]
}
\description{
Function definition. Resampling data by changing the sample rate from high frequency to low frequency.
}
\examples{
#Generate data frame
df <- data.frame("A"= runif(200, min=0, max=360),"B" = runif(200, min=0.1, max=10), "C" = rnorm(200))
df1 <- def.rsmp(data = df,FreqInp = 10,FreqOut = 5, MethRsmp = "filt",ColDisc = c(1))
df2 <- def.rsmp(data = df,FreqInp = 10,FreqOut = 0.5, MethRsmp = "zoo",ColDisc = c("A"))
}
\author{
Stefan Metzger \email{eddy4R.info@gmail.com}
Natchaya Pingintha-Durden
}
\references{
Currently none
License: Terms of use of the NEON FIU algorithm repository dated 2015-01-16
}
\seealso{
Currently none
}
\keyword{resampling}

