% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def.rtio.mass.h2o.dens.mole.R
\name{def.rtio.mass.h2o.dens.mole}
\alias{def.rtio.mass.h2o.dens.mole}
\title{Definition function: Calculation of the wet mass fraction (specifc humidity) from mole density of water vapor and mole density of dry air}
\usage{
def.rtio.mass.h2o.dens.mole(densMoleH2o, densMoleAirDry)
}
\arguments{
\item{\code{densMoleH2o}}{A vector containing mole density of water vapor, of class "numeric". [molH2o m-3]}

\item{\code{densMoleAirDry}}{A vector containing the mole density of dry air, of class "numeric". [mol m-3]}
}
\value{
The returned object is wet mass fraction (specific humidity)
}
\description{
Function definition. Calculate the wet mass fraction (specific humidity) of air mixture from mole density of water vapor and mole density of dry air
}
\examples{
Example 1, this will cause an error message due to densMoleH2o and densMoleAirDry have no units: 
def.rtio.mass.h2o.dens.mole(densMoleH2o = 0.3, densMoleAirDry = 41.1)
Example 2, assign values and units to variables first, the function should run ok.
densMoleH2o = 0.3
densMoleAirDry = 41.1
attributes(densMoleH2o)$unit = "molH2o m-3"
attributes(densMoleAirDry)$unit = "mol m-3"
def.rtio.mass.h2o.dens.mole (densMoleH2o, densMoleAirDry)
}
\references{
License: GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007.
}
\seealso{
Currently none.
}
\author{
Stefan Metzger \email{eddy4R.info@gmail.com}
Hongyan Luo \email{hluo@battelleecology.org}
}
\keyword{density}
\keyword{humidity,}
\keyword{mole}
\keyword{specific}
