% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def.rtio.mass.h2o.wet.pres.h2o.pres.R
\name{def.rtio.mass.h2o.wet.pres.h2o.pres}
\alias{def.rtio.mass.h2o.wet.pres.h2o.pres}
\title{Definition function: Calculate water vapor wet mass fraction (specific humidity) from water vapor pressure and static pressure}
\usage{
def.rtio.mass.h2o.wet.pres.h2o.pres(presH2o, pres)
}
\arguments{
\item{\code{presH2o}}{Either a vector or an object of class numeric of measured water vapor pressure and of the same length as \code{pres}. [Pa]}

\item{\code{pres}}{Either a vector or an object of class numeric of static pressure and of the same length as \code{presH2o}. [Pa]}
}
\value{
Specific humidity and of the same length as \code{presH2o} and \code{pres}. [kg kg-1]
}
\description{
Calculate specific humidity from water vapor pressure and static pressure.
}
\examples{
def.rtio.mass.h2o.wet.pres.h2o.pres(presH2o = 2212.04 , pres = 65000)
def.rtio.mass.h2o.wet.pres.h2o.pres(presH2o = c(2212.04, 348.69, 87.17) , pres = c(65000, 83000, 110000))
}
\author{
Stefan Metzger \email{eddy4R.info@gmail.com} \cr
Natchaya Pingintha-Durden
}
\references{
License: Terms of use of the NEON FIU algorithm repository dated 2015-01-16
}
\seealso{
Currently none
}
\keyword{conversion,}
\keyword{humidity}
\keyword{ratio}
\keyword{specific,}

