% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def.site.info.R
\name{def.site.info}
\alias{def.site.info}
\title{Definition function: Load pre-defined site specific information into the global environment}
\usage{
def.site.info(loc = c("IM", "LOS", "NR", "NS", "PF", "SERC", "CPER")[6])
}
\arguments{
\item{loc}{Parameter of class character. A parameter to choose the site for which site specific information is pulled.}
}
\value{
Site specific information provided as a list \code{SiteInfo}
}
\description{
Load pre-defined site specific information into the global environment.
}
\examples{
def.site.info("SERC")
}
\references{
License: GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007.
}
\seealso{
Currently none
}
\author{
Stefan Metzger \email{eddy4R.info@gmail.com}\cr
David Durden  \email{ddurden@battelleecology.org}
}
\keyword{metadata,}
\keyword{preprocessing}
\keyword{site,}
