% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def.smth.R
\name{def.smth}
\alias{def.smth}
\title{Data smoothing}
\usage{
def.smth(data, span, smthFunc = c("smthKern"))
}
\arguments{
\item{\code{data}}{A vector containing the input data. Of class "numeric" or "integer". []}

\item{\code{span}}{A Vector containing the span values. Of class "numeric" or "integer". The option \code{span} controls the degree of smoothing. For example; using span = c(15) to apply a long smoothing filter, or using spans = c(3,3) to apply two short filters in succession. [-]}

\item{\code{smthFunc}}{An object of class string containing the smooth functions ("smthKern"). []}
}
\value{
Estimated smooth values and of the same length as \code{data}. []
}
\description{
Function definition.Data smoothing method.
}
\examples{
def.smth(data = rnorm(100), span = c(15), smthFunc = c("smthKern"))
def.smth(data = rnorm(100), span = c(3,3))
}
\author{
Stefan Metzger \email{eddy4R.info@gmail.com} \cr
Natchaya Pingintha-Durden
}
\references{
License: Terms of use of the NEON FIU algorithm repository dated 2015-01-16 \cr
http://www.stat.rutgers.edu/home/rebecka/Stat565/lab42007.pdf \cr
}
\seealso{
\code{\link[stats]{kernel}}, \code{\link[stats]{kernapply}}
}
\keyword{Kernel}
\keyword{smoother}

