% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def.ssi.diff.R
\name{def.ssi.diff}
\alias{def.ssi.diff}
\title{Definition function: Calculate delta(or difference) of signal strength for LI-7200 IRGA}
\usage{
def.ssi.diff(ssiCo2, ssiH2o)
}
\arguments{
\item{\code{ssiCo2}}{A vector containing the CO2 signal strength, of class "numeric". [unitless]}

\item{\code{ssiH2o}}{A vector containing the H2O signal strength, of class "numeric". [unitless]}
}
\value{
The returned object is the signal strength difference between the CO2 signal strength and the H2O signal strength.
}
\description{
Function definition. Calculation of signal strength difference for LI-7200 IRGA.
}
\examples{
example 1 (This will give error message becuase ssiCo2 and ssiH2o have no units):
def.ssi.diff(ssiCo2 = 0.550, ssiH2o =0.561) 
example 2 (Assign the units and values to the variable before run function, which shoudl work fine.)
ssiCo2 =0.550
ssiH2o = 0.561 
attributes(ssiCo2)$unit <- "-"
attributes(ssiH2o)$unit <- "-"
def.ssi.diff(ssiCo2, ssiH2o)
}
\references{
License: GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007.
}
\seealso{
Currently none.
}
\author{
Stefan Metzger \email{eddy4R.info@gmail.com}
Hongyan Luo \email{hluo@battelleecology.org}
}
\keyword{irga}
\keyword{signal}
\keyword{strength,}
