% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def.temp.dew.pres.h2o.temp.mag.R
\name{def.temp.dew.pres.h2o.temp.mag}
\alias{def.temp.dew.pres.h2o.temp.mag}
\title{Definition function: Calculate dew point temperature from water vapor pressure and ambient temperature using Magnus equation}
\usage{
def.temp.dew.pres.h2o.temp.mag(presH2o, temp)
}
\arguments{
\item{\code{presH2o}}{Either a vector or an object of class numeric of measured water vapor pressure and of the same length as \code{temp}. [Pa]}

\item{\code{temp}}{Either a vector or an object of class numeric of measured air temperature and of the same length as \code{presH2o}. [K]}
}
\value{
Dew point temperature and of the same length as \code{presH2o} and \code{temp}. [K]
}
\description{
Calculate dew point temperature from water vapor pressure and ambient temperature using Magnus equation.
}
\examples{
def.temp.dew.pres.h2o.temp.mag(presH2o = 2212.04, temp = 298.15)
def.temp.dew.pres.h2o.temp.mag(presH2o = c(42.22, 87.22, 699.78), temp = c(265.15, 278.15, 293.15))
}
\author{
Stefan Metzger \email{eddy4R.info@gmail.com} \cr
Natchaya Pingintha-Durden \email{ndurden@battelleecology.org}
}
\references{
License: GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007.
Gueymard, C.: Assessment of the accuracy and computing speed of simplified saturation vapor equations using a new reference dataset, J. Appl. Meteorol., 32, 1294-1300, doi:10.1175/1520-0450(1993)0321294:aotaac2.0.co;2, 1993. \cr
}
\seealso{
Currently none
}
\keyword{Magnus}
\keyword{conversion,}
\keyword{dew}
\keyword{equation}
\keyword{humidity}
\keyword{point,}

