% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def.temp.mean.7200.R
\name{def.temp.mean.7200}
\alias{def.temp.mean.7200}
\title{Calculation of the average temperature in LI-7200 irga cell}
\usage{
def.temp.mean.7200(tempIn, tempOut, Val = FALSE)
}
\arguments{
\item{\code{tempIn}}{A vector containing the tempertaure measured at IRGA cell inlet, of class "numeric". [K]}

\item{\code{tempOut}}{A vector containing the tempertaure measured at IRGA cell outlet, of class "numeric". [K]}
}
\value{
The returned object is the the average temperature in LI-7200 IRGA cell calculated from the temperatrue at IRGA cell inlet and the temperature at IRGA cell outlet.
}
\description{
Function definition. Calculation of the average temperature in LI-7200 IRGA cell
}
\examples{
Example 1, this will cause an error message due to tempIn and tempOut have no units: 
def.temp.mean.7200(tempIn = 293, tempOut = 294)
Example 2, assign values and units to variables first, the function should run ok.
tempIn = 293
tempOut = 294
attributes(tempIn)$unit = "K"
attributes(tempOut)$unit = "K"
def.temp.mean.7200(tempIn, tempOut)
}
\author{
Stefan Metzger \email{eddy4R.info@gmail.com}
Hongyan Luo \email{eddy4R.info@gmail.com}
}
\references{
Currently none.
}
\seealso{
Currently none.
}
\keyword{irga}
\keyword{temperature,}

