% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def.temp.pres.pois.R
\name{def.temp.pres.pois}
\alias{def.temp.pres.pois}
\title{Poisson's equation (adiabatic change) - temperature as function of pressure change}
\usage{
def.temp.pres.pois(temp01, pres01, pres02,
  Kppa = eddy4R.base::IntlNatu$KppaDry)
}
\arguments{
\item{\code{pres01}}{Measured air pressure [same unit as reference pressure]}

\item{\code{pres02}}{Reference pressure [same unit as reference pressure]}

\item{\code{temp01}}{Measured air temperature [K]}

\item{\code{Kppa}}{Ratio of specific gas constant to specific heat at constant pressure. Default as KppaDry [-]}
}
\value{
Temperature at refernce pressure [K]
}
\description{
Poisson's equation (adiabatic change) - temperature as function of pressure change.
}
\examples{
temp02 <- def.temp.pres.pois(temp01 = 298.15, pres01 = 845, pres02 = 1000, Kppa = eddy4R.base::IntlNatu$KppaDry)
}
\author{
Stefan Metzger \email{eddy4R.info@gmail.com} \cr
Natchaya P-Durden
}
\references{
Currently none
}
\seealso{
Currently none
}
\keyword{Currently}
\keyword{none}

