% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def.temp.soni.R
\name{def.temp.soni}
\alias{def.temp.soni}
\title{Definition function: Calculation of sonic temperature from speed of sound}
\usage{
def.temp.soni(veloSoni)
}
\arguments{
\item{\code{veloSoni}}{A vector containing speed of sound, of class "numeric". [m s-1]}
}
\value{
The returned object is sonic temperature
}
\description{
Function definition. Calculation of sonic temperature from speed of sound using Eq. (9) in Appendix C of CSAT3 Three Dimensional Sonic Anemometer Instruction manual
}
\examples{
Example 1, this will cause an error message due to veloSoni has no units: 
def.temp.soni(veloSoni = 344)
Example 2, assign values and units to variables first, the function should run ok.
veloSoni = 344
attributes(veloSoni)$unit = "m s-1"
def.temp.soni(veloSoni)
}
\references{
License: GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007.
CSAT3 Three Dimensional Sonic Anemometer Instruction Mannual (Revisoin 2/15). Logan, Utah, USA. Campbell Scientifc, Inc. (2015)
}
\seealso{
Currently none.
}
\author{
Stefan Metzger \email{eddy4R.info@gmail.com}
Hongyan Luo \email{hluo@battelleecology.org}
}
\keyword{anemometer}
\keyword{sonic}
\keyword{temperature,}
