% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def.unit.intl.R
\name{def.unit.intl}
\alias{def.unit.intl}
\title{Find internal unit}
\usage{
def.unit.intl(unit)
}
\arguments{
\item{\code{unit}}{Required. A single character string providing the compount unit, constructed
with the rules specified in eddy4R.base::def.unit.info.R. \cr
#'}
}
\value{
A single character string specifying the corresponding eddy4R internal unit.
}
\description{
Function definition. Given an arbitraty input unit string, find the corresponding 
internal unit (string) used within the eddy4R family of functions.
}
\examples{
Currently none
}
\author{
Cove Sturtevant \email{eddy4R.info@gmail.com}
}
\references{
License: GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007.
eddy4R.base::def.unit.info.R
}
\seealso{
\code{\link{def.unit.info}}, \code{\link{IntlUnit$Intl}}, \code{\link{def.unit.conv}}
}
\keyword{base,}
\keyword{conversion,}
\keyword{input}
\keyword{internal}
\keyword{output}
\keyword{unit}
\keyword{units}
\keyword{units,}

