% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def.agr.vari.seSq.R
\name{def.vari.seSq.agr}
\alias{def.vari.seSq.agr}
\title{Determining mean, external, internal and total variance, and squared standard error}
\usage{
def.vari.seSq.agr(data = data.frame(mean, vari))
}
\arguments{
\item{\code{data}}{Dataframe of type numeric containing column vectors \code{mean} and \code{vari} of equal length.}

\item{\code{mean}}{Vector of type numeric. Means of the variable of interest at finer resolution, e.g. minutely [user-defined].}

\item{\code{vari}}{Vector of type numeric. Variances of the variable of interest at finer resolution, e.g. minutely [user-defined^2].}
}
\value{
Returns a list containing mean \code{mean}, external variance \code{variExt}, internal variance \code{variIntl}, total variance \code{variTota}, and squared standard error \code{seSq} at coarser resolution, e.g. hourly when \code{data} contains 60 minutely values.
}
\description{
Function defintion. Calculates the mean, external, internal and total variance, and the squared standard error over all supplied input values. Useful e.g. to aggregate from daily to monthly resolution by using all provided values, or similar.
}
\examples{
def.vari.seSq.agr(data = data.frame(mean = rnorm(10), vari = rnorm(10)^2))
}
\author{
Ke Xu \email{xuke2012abroad@gmail.com} \cr

Stefan Metzger \email{eddy4R.info@gmail.com}
}
\references{
license: Terms of use of the NEON FIU algorithm repository dated 2015-01-16
}
\seealso{
Currently none
}
\keyword{aggregate}
\keyword{deviation,}
\keyword{error,}
\keyword{standard}

