% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_site_info.r
\name{get_site_info}
\alias{get_site_info}
\title{Load pre-defined site location info into global environment}
\usage{
get_site_info(location = c("IM", "LOS", "NR", "NS", "PF", "SERC")[6])
}
\arguments{
\item{Currently}{none}
}
\value{
Currently none
}
\description{
Load pre-defined site location info into global environment.
}
\examples{
Currently none
}
\author{
Stefan Metzger \email{eddy4R.info@gmail.com}\cr
David Durden \email{eddy4R.info@gmail.com}
}
\references{
Currently none
}
\seealso{
Currently none
}
\keyword{Currently}
\keyword{none}

