% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smooth.r
\name{kernsmo}
\alias{kernsmo}
\title{Kernel smoother}
\usage{
kernsmo(data, spans)
}
\arguments{
\item{Currently}{none}
}
\value{
Currently none
}
\description{
Kernel smoother.
}
\examples{
Currently none
}
\author{
Stefan Metzger \email{eddy4R.info@gmail.com}
}
\references{
Currently none
}
\seealso{
Currently none
}
\keyword{Currently}
\keyword{none}

