% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.logger.singleton.R
\name{Logger.Singleton}
\alias{Logger.Singleton}
\title{Logger Class in Singleton Pattern}
\description{
A logger class in Singleton pattern with various logging levels
}
\examples{
# Create a logger instance
my_logger <- Logger$new(log_file = "logfile.txt")
# Create a logger instance without specifying a log file
my_logger <- Logger$new()
# Set the logging level to "debug"
my_logger$set_logging_level("debug")
# Log messages with debug level
my_logger$debug("This is a debug message.")
# Call the function with logging
some_function(my_logger)
# Get and print the log entries
log_entries <- my_logger$get_log_entries()
print("Log Entries:")
print(log_entries)
}
\section{Super class}{
\code{\link[R6P:Singleton]{R6P::Singleton}} -> \code{Logger.Singleton}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Logger.Singleton-new}{\code{Logger.Singleton$new()}}
\item \href{#method-Logger.Singleton-set_logging_level}{\code{Logger.Singleton$set_logging_level()}}
\item \href{#method-Logger.Singleton-set_log_file}{\code{Logger.Singleton$set_log_file()}}
\item \href{#method-Logger.Singleton-log_message}{\code{Logger.Singleton$log_message()}}
\item \href{#method-Logger.Singleton-debug}{\code{Logger.Singleton$debug()}}
\item \href{#method-Logger.Singleton-info}{\code{Logger.Singleton$info()}}
\item \href{#method-Logger.Singleton-warn}{\code{Logger.Singleton$warn()}}
\item \href{#method-Logger.Singleton-error}{\code{Logger.Singleton$error()}}
\item \href{#method-Logger.Singleton-get_log_entries}{\code{Logger.Singleton$get_log_entries()}}
\item \href{#method-Logger.Singleton-get_logging_level}{\code{Logger.Singleton$get_logging_level()}}
\item \href{#method-Logger.Singleton-get_log_file}{\code{Logger.Singleton$get_log_file()}}
\item \href{#method-Logger.Singleton-clone}{\code{Logger.Singleton$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Logger.Singleton-new"></a>}}
\if{latex}{\out{\hypertarget{method-Logger.Singleton-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Logger.Singleton$new(logging_level = NULL, log_file = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{logging_level}}{logging level}

\item{\code{log_file}}{file to store logging messages}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Logger.Singleton-set_logging_level"></a>}}
\if{latex}{\out{\hypertarget{method-Logger.Singleton-set_logging_level}{}}}
\subsection{Method \code{set_logging_level()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Logger.Singleton$set_logging_level(level)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{level}}{logging level}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Logger.Singleton-set_log_file"></a>}}
\if{latex}{\out{\hypertarget{method-Logger.Singleton-set_log_file}{}}}
\subsection{Method \code{set_log_file()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Logger.Singleton$set_log_file(file)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{file}}{logging file}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Logger.Singleton-log_message"></a>}}
\if{latex}{\out{\hypertarget{method-Logger.Singleton-log_message}{}}}
\subsection{Method \code{log_message()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Logger.Singleton$log_message(level, message)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{level}}{logging level}

\item{\code{message}}{logging message}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Logger.Singleton-debug"></a>}}
\if{latex}{\out{\hypertarget{method-Logger.Singleton-debug}{}}}
\subsection{Method \code{debug()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Logger.Singleton$debug(message)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{message}}{logging message}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Logger.Singleton-info"></a>}}
\if{latex}{\out{\hypertarget{method-Logger.Singleton-info}{}}}
\subsection{Method \code{info()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Logger.Singleton$info(message)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{message}}{logging message}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Logger.Singleton-warn"></a>}}
\if{latex}{\out{\hypertarget{method-Logger.Singleton-warn}{}}}
\subsection{Method \code{warn()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Logger.Singleton$warn(message)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{message}}{logging message}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Logger.Singleton-error"></a>}}
\if{latex}{\out{\hypertarget{method-Logger.Singleton-error}{}}}
\subsection{Method \code{error()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Logger.Singleton$error(message)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{message}}{logging message}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Logger.Singleton-get_log_entries"></a>}}
\if{latex}{\out{\hypertarget{method-Logger.Singleton-get_log_entries}{}}}
\subsection{Method \code{get_log_entries()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Logger.Singleton$get_log_entries()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Logger.Singleton-get_logging_level"></a>}}
\if{latex}{\out{\hypertarget{method-Logger.Singleton-get_logging_level}{}}}
\subsection{Method \code{get_logging_level()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Logger.Singleton$get_logging_level()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Logger.Singleton-get_log_file"></a>}}
\if{latex}{\out{\hypertarget{method-Logger.Singleton-get_log_file}{}}}
\subsection{Method \code{get_log_file()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Logger.Singleton$get_log_file()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Logger.Singleton-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Logger.Singleton-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Logger.Singleton$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
