% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maxcor.r
\name{maxcor}
\alias{maxcor}
\title{Lag two datasets, so as to maximise their correlations}
\usage{
maxcor(ref, obs, refDATA = ref, obsDATA = obs, obsVARS = NULL,
  maxlag = 2 * freq_loc, hardlag = FALSE, nplag = c("n", "p", "np")[3],
  absolute = FALSE, freq_loc = freq_res, hpf = TRUE, FRACmin = 0.1)
}
\arguments{
\item{Currently}{none}
}
\value{
Currently none
}
\description{
Lag two datasets, so as to maximise their correlations.
}
\examples{
Currently none
}
\author{
Stefan Metzger \email{eddy4R.info@gmail.com}
}
\references{
Currently none
}
\seealso{
Currently none
}
\keyword{Currently}
\keyword{none}

