% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RBPCdet.r
\name{RBPCdet}
\alias{RBPCdet}
\title{RMSD, bias, precision and coefficient of determination}
\usage{
RBPCdet(ref, test, perc = FALSE)
}
\arguments{
\item{Currently}{none}
}
\value{
Currently none
}
\description{
RMSD, bias, precision and coefficient of determination.
}
\examples{
Currently none
}
\author{
Stefan Metzger \email{eddy4R.info@gmail.com}
}
\references{
Currently none
}
\seealso{
Currently none
}
\keyword{Currently}
\keyword{none}

