% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RBPCdet.r
\name{RBPCdet2}
\alias{RBPCdet2}
\title{RMSD, bias, precision and coefficient of determination - incl. deadband}
\usage{
RBPCdet2(refe, test, perc = FALSE, db = NULL, db_rel = FALSE)
}
\arguments{
\item{Currently}{none}
}
\value{
Currently none
}
\description{
RMSD, bias, precision and coefficient of determination - incl. deadband.
}
\examples{
Currently none
}
\author{
Stefan Metzger \email{eddy4R.info@gmail.com}
}
\references{
Currently none
}
\seealso{
Currently none
}
\keyword{Currently}
\keyword{none}

