% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resamp.r
\name{resamp}
\alias{resamp}
\title{Resampling using rollapply() function}
\usage{
resamp(mat.raw, freq, freq_res, EIDAS)
}
\arguments{
\item{Currently}{none}
}
\value{
Currently none
}
\description{
Resampling using rollapply() function.
}
\examples{
Currently none
}
\author{
Stefan Metzger \email{eddy4R.info@gmail.com}
}
\references{
Currently none
}
\seealso{
Currently none
}
\keyword{Currently}
\keyword{none}

