% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resamp1.r
\name{resamp1}
\alias{resamp1}
\title{Resampling using zoo() function}
\usage{
resamp1(mat.raw, freq, freq_res, EIDAS)
}
\arguments{
\item{Currently}{none}
}
\value{
Currently none
}
\description{
Resampling using zoo() function.
}
\examples{
Currently none
}
\author{
Stefan Metzger \email{eddy4R.info@gmail.com}
}
\references{
Currently none
}
\seealso{
Currently none
}
\keyword{Currently}
\keyword{none}

