% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resamp1.r
\name{resamp2}
\alias{resamp2}
\title{Resampling using filter() function}
\usage{
resamp2(data, freq, freq_res, EIDAS)
}
\arguments{
\item{Currently}{none}
}
\value{
Currently none
}
\description{
Resampling using filter() function.
}
\examples{
Currently none
}
\author{
Stefan Metzger \email{eddy4R.info@gmail.com}
}
\references{
Currently none
}
\seealso{
Currently none
}
\keyword{Currently}
\keyword{none}

