% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spike.medfilt.r
\name{spike.medfilt}
\alias{spike.medfilt}
\title{Median filter de-spiking}
\usage{
spike.medfilt(DESP = list(dati, widt = 9, nbin = 2, rest = 10))
}
\arguments{
\item{Currently}{none}
}
\value{
Currently none
}
\description{
Median filter de-spiking.
}
\examples{
Currently none
}
\author{
Stefan Metzger \email{eddy4R.info@gmail.com}
}
\references{
Currently none
}
\seealso{
Currently none
}
\keyword{Currently}
\keyword{none}

