% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap.derv.prd.day.R
\name{wrap.derv.prd.day}
\alias{wrap.derv.prd.day}
\title{Wrapper function: Calculation of derived quantities (daily extent, native resolution)}
\usage{
wrap.derv.prd.day(data, ZoneTime)
}
\arguments{
\item{\code{data}}{List consisting of \code{ff::ffdf} file-backed objects, in the format provided by function \code{eddy4R.base::wrap.read.hdf5.neon.eddy()}. Of types numeric and integer.}

\item{\code{SiteLoca}}{List consisting of site-specific parameters. Of types numeric, integer and character.}
}
\value{
The returned object consistes of \code{data}, with the derived variables added to the respective list entry, and all list levels sorted alphabetically.
}
\description{
Wrapper function. Reads the list \code{data} in the format provided by function \code{eddy4R.base::wrap.read.hdf5.neon.eddy()}. For the list entries in \code{data} the following derived quantities are calculated, each through the call to a separate definition function: \cr
\code{data$time}: fractional UTC time, fractional day of year, local standard time;  \cr
\code{data$irga}: average signal strength, delta signal strength, total pressure, average temperature, water vapor partial pressure, water vapor saturation pressure, relative humidity, molar density of air (dry air and water vapor), molar density of dry air, wet mass fraction (specific humidity);  \cr
\code{data$soni}: sonic temperature.
}
\examples{
Currently none.
}
\author{
Stefan Metzger \email{eddy4R.info@gmail.com}
}
\references{
License: Terms of use of the NEON FIU algorithm repository dated 2015-01-16. \cr
}
\seealso{
Currently none.
}
\keyword{derived,}
\keyword{high-frequency,}
\keyword{irga,}
\keyword{post-processing,}
\keyword{pre-processing,}
\keyword{sonic,}
\keyword{time}

