% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap.hdf5.read.R
\name{wrap.hdf5.read}
\alias{wrap.hdf5.read}
\title{Wrapper function: Reading NEON HDF5 files}
\usage{
wrap.hdf5.read(DirInpLoca, SiteLoca, DateLoca, VarLoca, LvlTowr = c("000_040",
  "000_050", "000_060")[3], FreqLoca, Rglr = FALSE, Diag = FALSE,
  Rng = FALSE, RngLoca = NULL, DespLoca, MethMeas = c("ecte", "ecse")[1])
}
\arguments{
\item{DirInpLoca}{Character: Input directory.}

\item{SiteLoca}{Character: Site location.}

\item{DateLoca}{Character: Date in ISO format "(2016-05-26").}

\item{VarLoca}{Character: Which instrument to read data from.}

\item{LvlTowr}{The tower level that the sensor data is being collected in NEON data product convention (HOR_VER)}

\item{FreqLoca}{Integer: Measurement frequency.}

\item{RngLoca}{List of named ingegers: Thresholds for range test.}

\item{DespLoca}{List of integers: De-spiking parameters}

\item{MethMeas}{A vector of class "character" containing the name of measurement method (eddy-covariance turbulent exchange or storage exchange), MethMeas = c("ecte", "ecse"). Defaults to "ecte".}
}
\value{
Named list containing pre-processed time-series $time and $data.
}
\description{
Wrapper function. Reads an HDF5 input file in NEON standard format from \code{DirInpLoca}. Subsequently, (i) name and unit attributes are converted to eddy4R standard terms, (ii) variable units are converted accordingly, (iii) the data is regularized, (iv) sensor diagnostic tests are performed, (v) a range test is performed, and (vi) the data is de-spiked.
}
\examples{
Currently none.
}
\references{
License: GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007.
}
\seealso{
Currently none.
}
\author{
Stefan Metzger \email{eddy4R.info@gmail.com}
}
\keyword{conversion,}
\keyword{diagnostic,}
\keyword{file,}
\keyword{pre-processing,}
\keyword{range,}
\keyword{read,}
\keyword{regularization,}
\keyword{spike}
\keyword{unit}
