% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap.hdf5.wrte.dp01.api.R
\name{wrap.hdf5.wrte.dp01.api}
\alias{wrap.hdf5.wrte.dp01.api}
\title{Wrapper function: Gather/Write reingested NEON Level 1 data, qfqm, and uncertainty to output HDF5}
\usage{
wrap.hdf5.wrte.dp01.api(date, FileOut, SiteLoca, DpName = c("tempAirLvl",
  "tempAirTop"), LvlTowr, TimeAgr = c(1, 30))
}
\arguments{
\item{date}{Character: The date for the data to be gathered in ISO format ("YYYYmmdd").}

\item{FileOut}{Character: The file name for the output HDF5 file}

\item{SiteLoca}{Character: Site location.}

\item{LvlTowr}{Character: The tower level that the sensor data is being collected in NEON data product convention (HOR_VER).}

\item{TimeAgr}{Integer: The time aggregation index in minutes (i.e. 30)}

\item{Dp01}{Character: A vector of data product names for the data to be gathered.}
}
\value{
An updated dp0p HDF5 file with dp01 data, qfqm, and uncertainty written
}
\description{
Definition function. To write NEON Level 1 data product descriptive statistics (mean, minimum, maximum, variance, number of non-NA points), quality flags and quality metrics, and uncertainty values gathered from via the API to an output HDF5 file.
}
\examples{
Currently none.
}
\references{
License: GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007.
}
\seealso{
Currently none.
}
\author{
David Durden \email{ddurden@battelleecology.org}
}
\keyword{HDF5}
\keyword{output,}
