% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap.hdf5.wrte.dp01.R
\name{wrap.hdf5.wrte.dp01}
\alias{wrap.hdf5.wrte.dp01}
\title{Wrapper function: Write NEON Level 1 data, qfqm, and uncertainty to output HDF5}
\usage{
wrap.hdf5.wrte.dp01(
  inpList,
  FileInp,
  FileOut,
  SiteLoca,
  LvlTowr,
  MethUcrt = TRUE,
  MethDp04 = FALSE,
  MethSubAgr = TRUE
)
}
\arguments{
\item{inpList}{A list of including dp01 data, quality flags and quality metrics, and uncertainty calculations to package and write to an output HDF5 file}

\item{FileInp}{The file name for the input dp0p HDF5 file to grab metadata}

\item{FileOut}{The file name for the output HDF5 file}

\item{SiteLoca}{Character: Site location.}

\item{LvlTowr}{The tower level that the sensor data is being collected in NEON data product convention (HOR_VER)}

\item{MethUcrt}{Logical: Determines if uncertainty information is available for output.}

\item{MethDp04}{logical indicating if ECTE dp04 HDF5 data should be included.}

\item{MethSubAgr}{Logical: Determines if 1-minute data is available for output.}
}
\value{
An HDF5 file with dp01 data, qfqm, and uncertainty written
}
\description{
Wrapper function. To write NEON Level 1 data product descriptive statistics (mean, minimum, maximum, variance, number of non-NA points), quality flags and quality metrics, and uncertainty quantification to an output HDF5 file.
}
\examples{
Currently none.
}
\references{
License: GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007.
}
\seealso{
Currently none.
}
\author{
David Durden \email{ddurden@battelleecology.org}
}
\keyword{HDF5}
\keyword{output,}
