% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap.irga.vali.R
\name{wrap.irga.vali}
\alias{wrap.irga.vali}
\title{Wrapper function: Validation processing for IRGA}
\usage{
wrap.irga.vali(data, qfqmFlag, gasRefe, DateProc)
}
\arguments{
\item{data}{List consisting of \code{ff::ffdf} file-backed objects containing the dp0p input IRGA.}

\item{qfqmFlag}{List consisting of \code{ff::ffdf} file-backed objects containing the IRGA quality flags.}

\item{gasRefe}{List containing the values of the reference gases. [mol mol-1]}

\item{DateProc}{A vector of class "character" containing the processing date.}
}
\value{
The returned object consists of:\cr
\code{rtioMoleDryCo2Vali} scriptive statistics (mean, min, max, vari, numSamp, se) of CO2 dry mole concentration during performing validation and CO2 dry mole concentration of reference gases.\cr
\code{rtioMoleDryCo2Mlf} linear regression coefficients resulted from the maximum-likelihood fitting of a functional relationship. \cr
\code{rtioMoleDryCo2Cor} dataframe consists of the correction IRGA sub data products.
}
\description{
Wrapper function to apply IRGA validation.
}
\examples{
Currently none.
}
\references{
License: GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007.
}
\seealso{
Currently none.
}
\author{
Natchaya P-Durden \email{eddy4R.info@gmail.com}
}
\keyword{derived,}
\keyword{irgaTurb,}
\keyword{post-processing,}
\keyword{pre-processing,}
\keyword{validation}
