% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap.neon.dp01.agr.prd.R
\name{wrap.neon.dp01.agr.prd}
\alias{wrap.neon.dp01.agr.prd}
\title{Wrapper function: Create NEON Level 1 data products with quality flags and quality metrics for different aggregation periods}
\usage{
wrap.neon.dp01.agr.prd(inpList)
}
\arguments{
\item{inpList}{A list of including dp0p data and quality flags to perform dp01 calculation and quality tests.}
}
\value{
A list containing dp01 data and quality flags and metrics for all sensors including the following: \cr
\code{data}:
\code{mean} The mean of non-NA values in \code{data}
\code{min} The minimum value of non-NA values in \code{data}
\code{max} The maximum value of non-NA values in \code{data}
\code{vari} The variance of non-NA values in \code{data}
\code{num} The number of non-NA values in \code{data}
\code{se} The standard error of the mean of non-NA values in \code{data}
\code{qfqm}: 
\code{qm}  A list of data frame's containing quality metrics (fractions) of failed, pass, and NA for each of the individual flag which related to L1 sub-data products if RptExpd = TRUE. [fraction] \cr
\code{qmAlph} A dataframe containing metrics in a  columns of class "numeric" containing the alpha quality metric for L1 sub-data products. [fraction] \cr
\code{qmBeta} A dataframe containing metrics in a columns of class "numeric" containing the beta quality metric for L1 sub-data products. [fraction] \cr
\code{qfFinl} A dataframe containing flags in a columns of class "numeric", [0,1], containing the final quality flag for L1 sub-data products. [-] \cr
\code{qfSciRevw} A dataframe containing flags in a columns of class "numeric", [0,1], containing the scientific review quality flag for L1 sub-data products. [-] \cr
}
\description{
Wrapper function. Compute NEON Level 1 data products with quality flags and quality metrics for different aggregation periods (e.g. 1 minute).
}
\examples{
Currently none
}
\references{
License: GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007.

License: GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007.
}
\seealso{
Currently none.
}
\author{
David Durden \email{ddurden@battelleecology.org}
}
\keyword{NEON}
\keyword{QAQC,}
\keyword{aggregate,}
\keyword{average,}
\keyword{descriptive}
\keyword{flags,}
\keyword{metrics}
\keyword{quality}
\keyword{statistics,}
