% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap.neon.dp01.R
\name{wrap.neon.dp01}
\alias{wrap.neon.dp01}
\title{Create NEON Level 1 data product descriptive statistics across list elements}
\usage{
wrap.neon.dp01(data, idx = NULL)
}
\arguments{
\item{\code{data}}{A data.frame or list containing the L0p (calibrated) input data at native resolution. Of type numeric or integer. [-]}

\item{\code{idx}}{If data is a list, which list entries should be processed into Level 1 data products? Defaults to NULL which expects data to be a data.frame. Of type character. [-]}
}
\value{
Descriptive statistics, for \code{vrbs = FALSE} a data frame and for \code{vrbs = TRUE} a list:\cr
\code{mean} The mean of non-NA values in \code{data}
\code{min} The minimum value of non-NA values in \code{data}
\code{max} The maximum value of non-NA values in \code{data}
\code{vari} The variance of non-NA values in \code{data}
\code{num} The number of non-NA values in \code{data}
\code{se} The standard error of the mean of non-NA values in \code{data}
}
\description{
Wrapper function. Compute NEON Level 1 data product descriptive statistics (mean, minimum, maximum, variance, number of non-NA points) across list elements.
}
\examples{
  # data.frame which variables contain the unit attributes
  
    # create list with L0p data
    data <- list(
      sens01 = data.frame(
        velo = rnorm(10),
        temp = rnorm(10)),
      sens02 = data.frame(
        velo = rnorm(10),
        temp = rnorm(10)),
      sens03 = data.frame(
        velo = rnorm(10),
        temp = rnorm(10))
      )
  
    # assign unit attribute
    attributes(data$sens01$velo)$unit <- "m s-1"
    attributes(data$sens01$temp)$unit <- "K"
    attributes(data$sens02$velo)$unit <- "m s-1"
    attributes(data$sens02$temp)$unit <- "K"
    attributes(data$sens03$velo)$unit <- "m s-1"
    attributes(data$sens03$temp)$unit <- "K"
  
  # calculate L1 data products only for sensors sens01 and sens02
  rpt <- wrap.neon.dp01(data = data, idx = c("sens01", "sens02"))
  # units are propagated
  attributes(rpt$sens02$se$velo)$unit
  # [1] "m s-1"
}
\author{
Stefan Metzger \email{eddy4R.info@gmail.com} \cr
}
\references{
Currently none.
}
\seealso{
Currently none.
}
\keyword{aggregate,}
\keyword{average,}
\keyword{descriptive}
\keyword{statistics}

