% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap.para.thrs.R
\name{wrap.para.thsh}
\alias{wrap.para.thsh}
\title{Definition Function to read threshold table from CI-Parameter-Repo}
\usage{
wrap.para.thsh(DpName = c("IrgaTurb", "MfcSampTurb", "Soni", "Amrs"),
  DirInp = "~/eddy/data/Thresholds_EC/CI-Parameter-Repo/ParaSci/Ecte/Dp0p",
  DirOut = "~/eddy/data/Thresholds_EC/threshold_ecte", FileWrte = FALSE)
}
\value{
A data.frame consisting of the threshold values to be used for the provided site.
}
\description{
Workflow to read ECSE threshold table from CI-Parameter-Repo
}
\seealso{
Currently none
}
\author{
Natchaya Pingintha-Durden \email{ndurden@battelleecology.org}
David Durden \email{ddurden@battelleecology.org}
}
