% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap.prd.day.ecse.R
\name{wrap.prd.day.ecse}
\alias{wrap.prd.day.ecse}
\title{Wrapper function: To perform daily ECSE processing in native resolution}
\usage{
wrap.prd.day.ecse(inpList)
}
\arguments{
\item{inpList}{List consisting of input data in the format provided by function \code{eddy4R.base::wrap.neon.read.hdf5.eddy()}. Of types numeric and integer.}
}
\value{
The returned object consistes of \code{rpt} after applying the daily processing.
}
\description{
Wrapper function. To perform daily ECSE processing in native resolution
}
\examples{
Currently none.
}
\references{
License: GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007.
}
\seealso{
Currently none.
}
\author{
Natchaya Pingintha-Durden \email{eddy4R.info@gmail.com}
}
\keyword{ECSE,}
\keyword{daily}
\keyword{processing,}
\keyword{qfqm,}
\keyword{quality}
