% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap.read.hdf5.neon.eddy.R
\name{wrap.read.hdf5.neon.eddy}
\alias{wrap.read.hdf5.neon.eddy}
\title{Reading NEON HDF5 files}
\usage{
wrap.read.hdf5.neon.eddy(DirInpLoca, SiteLoca, DateLoca, VarLoca, FreqLoca,
  Rglr = FALSE, Diag = FALSE, Rng = FALSE, RngLoca, DespLoca)
}
\arguments{
\item{\code{DirInpLoca}}{Character: Input directory.}

\item{\code{SiteLoca}}{Character: Site location.}

\item{\code{DateLoca}}{Character: Date in ISO format "(2016-05-26").}

\item{\code{VarLoca}}{Character: Which instrument to read data from.}

\item{\code{FreqLoca}}{Integer: Measurement frequency.}

\item{\code{RngLoca}}{List of named ingegers: Thresholds for range test.}

\item{\code{DespLoca}}{List of integers: De-spiking parameters}
}
\value{
Named list containing pre-processed time-series $time and $data.
}
\description{
Wrapper function. Reads an HDF5 input file in NEON standard format from \code{DirInpLoca}. Subsequently, (i) name and unit attributes are converted to eddy4R standard terms, (ii) variable units are converted accordingly, (iii) the data is regularized, (iv) sensor diagnostic tests are performed, (v) a range test is performed, and (vi) the data is de-spiked.
}
\examples{
Currently none.
}
\author{
Stefan Metzger \email{eddy4R.info@gmail.com}
}
\references{
License: Terms of use of the NEON FIU algorithm repository dated 2015-01-16. \cr
}
\seealso{
Currently none.
}
\keyword{conversion,}
\keyword{diagnostic,}
\keyword{file,}
\keyword{pre-processing,}
\keyword{range,}
\keyword{read,}
\keyword{regularization,}
\keyword{spike}
\keyword{unit}

