% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap.time.rglr.dp00.ecse.R
\name{wrap.time.rglr.dp00.ecse}
\alias{wrap.time.rglr.dp00.ecse}
\title{Wrapper function: Time regularization for ECSE dp00}
\usage{
wrap.time.rglr.dp00.ecse(DirIn, Date, Site = "CPER", Dom = "D10", Freq,
  IdDp00, horVer)
}
\arguments{
\item{\code{DirIn}}{Character: Input directory. [-]}

\item{\code{Date}}{Character: Processing date e.g. "20170521". [-]}

\item{\code{Site}}{Character: Site location. [-]}

\item{\code{Dom}}{Character: Domain. [-]}

\item{\code{Freq}}{Desired frequency of  the regularized dataset. Of class "numeric" or "integer". [Hz]}

\item{\code{IdDp00}}{Character: dp00 data product number. [-]}

\item{\code{horVer}}{Character: Horizontal and vertical location of dp00. [-]}
}
\value{
A dataframe including the regularized dp00. [User-defined]
}
\description{
Wrapper function. Time regularization for ECSE dp00
}
\examples{
Currently none.
}
\references{
License: Terms of use of the NEON FIU algorithm repository dated 2015-01-16. \cr
}
\seealso{
Currently none.
}
\author{
Natchaya Pingintha-Durden \email{ndurden@battelleecology.org}
}
\keyword{ECSE}
\keyword{L0,}
\keyword{NEON,}
\keyword{dp00,}
\keyword{eddy-covariance,}
\keyword{flux,}
\keyword{storage}
