% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap.unit.conv.out.ec.R
\name{wrap.unit.conv.out.ec}
\alias{wrap.unit.conv.out.ec}
\title{Wrapper function: Output unit conversion for ECTE}
\usage{
wrap.unit.conv.out.ec(inpList, MethType = c("Data", "Ucrt", "Qfqm",
  "Vali")[1])
}
\arguments{
\item{inpList}{Required. A named list of data frames of type numeric, containing the data to be converted.}

\item{MethType}{Required. A character string containing the type of data to be converted. Defauts to \code{MethType} = c("Data").}
}
\value{
A list, \code{rpt}, with data, qfqm, or uncertainty output with the correct output units
}
\description{
Function wrapper. Convert a list of data to eddy4r output units using def.unit.conv function, with special attention to variable with the mean removed that are translated between units (i.e. variance for temperature when converting K to C).
}
\examples{
Currently none.
}
\references{
License: GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007.
}
\seealso{
Currently none.
}
\author{
David Durden \email{eddy4R.info@gmail.com}
}
\keyword{HDF5}
\keyword{output,}
