% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def.bin.R
\name{def.bin}
\alias{def.bin}
\title{Binning data}
\usage{
def.bin(idep, depe, RngMinMax = NULL, NumBin, widtBin = c("lin", "log10",
  "exp10", "logExp", "expLog"), meanFunc = c("mean", "median"))
}
\arguments{
\item{\code{idep}}{Either a vector or matrix of class numeric or integer containing the independent variable and of the same length as \code{depe}. []}

\item{\code{depe}}{Either a vector or matrix of class numeric or integer containing the dependent variable and of the same length as \code{idep}. []}

\item{\code{RngMinMax}}{An object of class numeric or integer containing the minimum and maximum values of the independent variable. Defaults to NULL. []}

\item{\code{NumBin}}{An object of class numeric or integer containing the number of bins. []}

\item{\code{widtBin}}{An object of class string containing the functions ("lin", "log10", "exp10", "logExp", "expLog") to determine bin width distribution of the independent variable. []}

\item{\code{meanFunc}}{An object of class string containing the arithmetic "mean" and "median". []}
}
\value{
\code{idep} A list object of class "numeric" containing the resulted binning of independent variable and of the same length as {widtBin} and \code{depe} a matrix containing the the resulted binning of dependent variable and of the same length as {widtBin}. \cr
}
\description{
Function definition. Smooth data using Binning method.
}
\examples{
def.bin(idep = rnorm(5000), depe = rnorm(5000), RngMinMax = NULL, NumBin = 23, widtBin = "log10", meanFunc = "mean" )
def.bin(idep = rnorm(500), depe = rnorm(500), RngMinMax = c(0.1, 0.4), NumBin = 12, widtBin = "lin", meanFunc = "median" )
}
\author{
Stefan Metzger \email{eddy4R.info@gmail.com} \cr
Natchaya Pingintha-Durden
}
\references{
License: Terms of use of the NEON FIU algorithm repository dated 2015-01-16
}
\seealso{
Currently none
}
\keyword{bin}

