% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def.cart.pol.R
\name{def.cart.pol}
\alias{def.cart.pol}
\title{Definition function: Decomposing azimuth angles to cartesian vectors}
\usage{
def.cart.pol(az)
}
\arguments{
\item{\code{az}}{vector of type numeric, clockwise azimuth angle with 0 / 360 degree discontinuity in north [decimal degrees]}
}
\value{
matrix consisting of unit vectors X (1st column, -1 ... 1, positive to north) and Y (2nd column, -1 ... 1, positive to east)
}
\description{
Decompose azimuth angles from polar to cartesian for angular averaging.
}
\examples{
def.cart.pol(20)
}
\author{
Stefan Metzger \email{eddy4R.info@gmail.com} \cr
Natchaya P-Durden
}
\references{
Currently none
}
\seealso{
Currently none
}
\keyword{unit}
\keyword{vector}

