% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def.coef.corl.R
\name{def.coef.corl}
\alias{def.coef.corl}
\title{Definition function: Coriolis coefficient}
\usage{
def.coef.corl(lat)
}
\arguments{
\item{\code{lat}}{Required. Latitude [degrees North]}
}
\value{
Coriolis coefficient for \code{lat} [rad s-1].
}
\description{
Function definition. Determine coriolis coefficient.
}
\examples{
Currently none
}
\author{
Stefan Metzger \email{eddy4R.info@gmail.com} \cr
Cove Sturtevant \email{eddy4R.info@gmail.com}
}
\references{
Currently none
}
\seealso{
Currently none
}
\keyword{constants}
\keyword{natural}

