% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def.dens.mass.h2o.press.h2o.temp.R
\name{def.dens.mass.h2o.press.h2o.temp}
\alias{def.dens.mass.h2o.press.h2o.temp}
\title{Definition function: Calculate absolute humidity from water vapor pressure and ambient temperature}
\usage{
def.dens.mass.h2o.press.h2o.temp(presH2o, temp)
}
\arguments{
\item{presH2o}{Either a vector or an object of class numeric of measured water vapor pressure and of the same length as \code{temp}. [Pa]}

\item{temp}{Either a vector or an object of class numeric of measured air temperature and of the same length as \code{presH2o}. [K]}
}
\value{
Absolute humidity and of the same length as \code{presH2o} and \code{temp}. [kg m-3]
}
\description{
Calculate absolute humidity from water vapor pressure and ambient temperature.
}
\examples{
def.dens.mass.h2o.press.h2o.temp(presH2o = 2054.04, temp = 298.15)
def.dens.mass.h2o.press.h2o.temp(presH2o = c(42.22, 348.70, 2054.04), temp = c(265.15, 278.15, 298.15))
}
\references{
License: GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007
}
\seealso{
Currently none
}
\author{
Stefan Metzger \email{eddy4R.info@gmail.com} \cr
Natchaya Pingintha-Durden \email{ndurden@battelleecology.org}
}
\keyword{conversion}
\keyword{humidity}
