% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def.dld.zip.R
\name{def.dld.zip}
\alias{def.dld.zip}
\title{Definition function: Download and extract .zip archives from a web address}
\usage{
def.dld.zip(Inp = list(Url, Dir = tempdir()))
}
\arguments{
\item{Inp}{The input parameter list containing \code{Url} and \code{Dir}.}

\item{Url}{The internet address, of class "character". For use with Dropbox, Google Drive, OneNote... the \href{https://www.google.com/search?q=Direct+Link+Generator}{direct download link} needs to be used. For example in case of Dropbox the web address needs to end on ...dl=1.}

\item{Dir}{The target directory, of class "character".}
}
\value{
The function saves the extracted .zip archive into the directory \code{Dir}. Currently no values are returned to the environment it is called from.
}
\description{
Function defintion. Downloads .zip archives from a web \code{Url}, and saves and extracts the archive into the directory \code{Dir}.
}
\examples{
def.dld.zip(Inp = list(
Url = "https://www.dropbox.com/s/1yv1ais1wdvoq1l/outRefe_20160410.zip?dl=1",
Dir = tempdir()))
}
\references{
License: GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007. \cr
\url{http://hydroecology.net/downloading-extracting-and-reading-files-in-r/}. \cr
}
\seealso{
Currently none.
}
\author{
Stefan Metzger \email{eddy4R.info@gmail.com}
}
\keyword{Dropbox,}
\keyword{archive,}
\keyword{download}
\keyword{zip,}
