% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def.extr.hdf5.R
\name{def.extr.hdf5}
\alias{def.extr.hdf5}
\title{Definition function: Extract data and attributes from HDF5 input file to another HDF5 output file}
\usage{
def.extr.hdf5(FileIn, FileOut = NULL, MethExtrData = TRUE,
  MethExtrAttr = TRUE, dp01 = NULL)
}
\arguments{
\item{FileIn}{is the input HDF5 file (turb or stor) the data and metadata are being read from.}

\item{FileOut}{is the output file nsae HDF5 written to.}

\item{MethExtrData}{logical parameter that decides if data from the input file should be extracted and written to the output file.}

\item{MethExtrAttr}{logical parameter that decides if attributes (metadata) from the input file should be extracted and written to the output file.}

\item{dp01}{A vector of class "character" containing the name of NEON dp01 which data from the input file are not extracted and written to the output file}
}
\value{
A NEON formatted HDF5 file that has parameters from the input file written to the output HDF5 file.
}
\description{
Definition function. Function extracts group sturcture, data, and metadata attributes from input, examples include ecte (turbulent/turb) and ecse (storage/stor) HDF5 files to a another HDF5 file, used for nsae (net surface atmosphere exchange), with the same group heirarchy structure from each file.
}
\examples{
Currently none
}
\references{
License: GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007. \cr
NEON Algorithm Theoretical Basis Document:Eddy Covariance Turbulent Exchange Subsystem Level 0 to Level 0’ data product conversions and calculations (NEON.DOC.000823) \cr
NEON Algorithm Theoretical Basis Document:Eddy Covariance Turbulent Exchange Subsystem Level 1 data product calculations (NEON.DOC.000807)
}
\seealso{
Currently none
}
\author{
David Durden \email{ddurden@battelleecology.org}
}
\keyword{ECSE,}
\keyword{ECTE,}
\keyword{HDF5,}
\keyword{NEON,}
\keyword{eddy-covariance,}
\keyword{stor}
\keyword{turb,}
