% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def.hdf5.crte.R
\name{def.hdf5.crte}
\alias{def.hdf5.crte}
\title{Definition function: Create the ECTE HDF5 file structure}
\usage{
def.hdf5.crte(Date, Site = "SERC", LevlTowr, DirOut, Dom = Dom,
  MethExpd = TRUE, fileNameReadMe = NULL, fileNameObjDesc = NULL)
}
\arguments{
\item{Date}{is the date for the output file being generated.}

\item{Site}{is the site for which the output file is being generated.}

\item{LevlTowr}{is the measurement level of the tower top to determine the VER number of the NEON DP naming convention.}

\item{DirOut}{is the output directory where the file being generated is stored.}

\item{Dom}{is the NEON domain}

\item{MethExpd}{logical indicating if the output should be expanded or basic}

\item{fileNameReadMe}{character indicating the filename incl. absolute path to the ReadMe file for inclusion in the output HDF5 file. Defaults to \code{NULL}, which downloads the readme file from a web location.}

\item{fileNameObjDesc}{character indicating the filename incl. absolute path to the object description file for inclusion in the output HDF5 file. Defaults to \code{NULL}, which downloads the object description file from a web location.}
}
\value{
A NEON formatted HDF5 file that is output to /code{DirOut} with a readme and object description included.
}
\description{
Definition function. Function creates the standard NEON HDF5 file structure for the ECTE data products.
}
\examples{
#Example run to create L0 gold files for 4/24/2016 at SERC
#Setting Site
Site <- "SERC"
LevlTowr <- "000_060"
Dom <- "D02"
MethExpd <- TRUE
#Setting Date to be processed
Date <- "2016-04-24"
#Set directory for example output to working directory
DirOut <- getwd()
#Running example
def.hdf5.crte(Date = Date, Site = Site, LevlTowr = LevlTowr, DirOut = DirOut, Dom = Dom, MethExpd = MethExpd)
}
\references{
License: GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007. \cr
NEON Algorithm Theoretical Basis Document:Eddy Covariance Turbulent Exchange Subsystem Level 0 to Level 0’ data product conversions and calculations (NEON.DOC.000807)
}
\seealso{
Currently none
}
\author{
David Durden \email{ddurden@battelleecology.org} \cr
Stefan Metzger \email{eddy4R.info@gmail.com}
}
\keyword{ECTE}
\keyword{HDF5,}
\keyword{NEON,}
\keyword{eddy-covariance,}
