% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def.hdf5.str.R
\name{def.hdf5.crte}
\alias{def.hdf5.crte}
\title{Definition function: Create the ECTE HDF5 file structure}
\usage{
def.hdf5.crte(Date, Site = "SERC", LevlTowr, DirOut, LevlDp = "dp01")
}
\arguments{
\item{\code{Date}}{is the date for the output file being generated.}

\item{\code{Site}}{is the site for which the output file is being generated.}

\item{\code{LevlTowr}}{is the measurement level of the tower top to determine the VER number of the NEON DP naming convention.}

\item{\code{DirOut}}{is the output directory where the file being generated is stored.}

\item{\code{LevlDp}}{is output file DP level for the file naming.}
}
\value{
A NEON formatted HDF5 file that is output to /code{DirOut}
}
\description{
Definition function. Function creates the standard NEON HDF5 file structure for the ECTE data products.
}
\examples{
#Example run to create L0 gold files for 4/24/2016 at SERC
#Setting Site
Site <- "SERC"
LevlTowr <- "000_060"
LevlDp <- "dp01"
#Setting Date to be processed
Date <- "2016-04-24"
#Set directory for example output to working directory
DirOut <- getwd()
#Running example
def.hdf5.crte(Date = Date, Site = Site, LevlTowr = LevlTowr, DirOut = DirOut, LevlDp)
}
\author{
Dave Durden \email{ddurden@battelleecology.org}
}
\references{
Currently none.
}
\seealso{
Currently none
}
\keyword{ECTE}
\keyword{HDF5,}
\keyword{NEON,}
\keyword{eddy-covariance,}

