% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def.irga.vali.cor.R
\name{def.irga.vali.cor}
\alias{def.irga.vali.cor}
\title{definition function: Calculation of time-series of the correction IRGA sub data products.}
\usage{
def.irga.vali.cor(
  data,
  DateProc,
  coef,
  valiData,
  valiCrit = FALSE,
  ScalMax = FALSE,
  FracSlp = FALSE,
  OfstMax = FALSE,
  Freq = 20
)
}
\arguments{
\item{data}{List consisting of \code{ff::ffdf} file-backed objects containing the dp0p input IRGA.}

\item{DateProc}{A vector of class "character" containing the processing date.}

\item{coef}{List consists of linear regression coefficients (slope and offset) for DateProc - 1, DateProc, and DateProc + 1.}

\item{valiData}{List consisting of descriptive statistics (mean, min, max, vari, numSamp, se) of CO2 dry mole concentration during performing validation and CO2 dry mole concentration of reference gases.}

\item{valiCrit}{A logical stating if there are more than one validation occurred within DateProc.}

\item{ScalMax}{Maximum scale value. The validation correction will not apply if scale (resulted from maximum-likelihood fitting of a functional relationship (MLFR)) is greater than ScalMax or ScalMax = FALSE. Defaults to FALSE.}

\item{FracSlp}{Upper and lower bounds of slope values. The validation correction will not apply if slope (resulted from regression fitting) is greater/lower than the FracSlp maximum or minimum value or FracSlp = FALSE. Defaults to FALSE.}

\item{OfstMax}{Maximum offset value. The validation correction will not apply if slope (resulted from regression fitting) is greater than the OfstMax (unit in mol mol-1) or OfstMax = FALSE. Defaults to FALSE.}

\item{Freq}{Measurement frequency. Defaults to 20. [Hz]}
}
\value{
The returned dataframe consists of the correction IRGA sub data products.
}
\description{
Definition function to calculate time-series of the correction IRGA sub data products based on linear interpolation method.
}
\examples{
Currently none.
}
\references{
License: GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007.
}
\seealso{
Currently none.
}
\author{
Natchaya P-Durden \email{eddy4R.info@gmail.com}
}
\keyword{derived,}
\keyword{irgaTurb,}
\keyword{post-processing,}
\keyword{pre-processing,}
\keyword{validation}
