% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def.irga.vali.thsh.R
\name{def.irga.vali.thsh}
\alias{def.irga.vali.thsh}
\title{Definition function: Threshold IRGA validation data based on benchmarking regression}
\usage{
def.irga.vali.thsh(data, DateProc, bnchSlpMax, bnchSlpMin)
}
\arguments{
\item{data}{List of validation data as a report from eddy4R.base::wrap.irga.vali()}

\item{DateProc}{A vector of class "character" containing the processing date.}

\item{bnchSlpMax}{Maximum acceptable slope of the benchmarking regression}

\item{bnchSlpMin}{Minimum acceptable slope of the benchmarking regression}

\item{gasRefe}{List containing the values of the reference gases. [mol mol-1]}
}
\value{
pass/fail criteria for a day's validation
}
\description{
Wrapper function to apply IRGA validation.
}
\examples{
Currently none.
}
\references{
License: GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007.
}
\seealso{
Currently none.
}
\author{
Chris Florian \email{eddy4R.info@gmail.com}
}
\keyword{derived,}
\keyword{irgaTurb,}
\keyword{post-processing,}
\keyword{pre-processing,}
\keyword{validation}
