% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def.lag.r
\name{def.lag}
\alias{def.lag}
\title{Definition function: Lag two datasets, so as to maximise their cross-correlation}
\usage{
def.lag(ref, obs, freq_loc, refDATA = ref, obsDATA = obs, obsVARS = NULL,
  maxlag = 2 * freq_loc, hardlag = TRUE, nplag = c("n", "p", "np")[3],
  absolute = TRUE, hpf = TRUE, FRACmin = 0.1)
}
\arguments{
\item{\code{ref}}{A vector with variable in reference time frame. Of class numeric. [-]}

\item{\code{obs}}{A vector with variable in time frame to be adjusted. Of class numeric. [-]}

\item{\code{freq_loc}}{Acquisition frequency of ref and obs. Of class ingeter. [Hz]}

\item{\code{refDATA}}{A matrix or data.frame with all data that carries the time frame of ref. Defaults to ref. Of any class. [-]}

\item{\code{obsDATA}}{A matrix or data.frame with all data that carries the time frame of obs. Defaults to obs. Of any class. [-]}

\item{\code{obsVARS}}{A vector specifying if only several columns in obsDATA shall be lagged. Defaults to NULL. Of class integer or character. [-]}

\item{\code{maxlag}}{Maximum lag, by default 2 x freq_loc. Of class integer. [-]}

\item{\code{hardlag}}{TRUE - interpret maxlag as maximum permissible lag; FALSE - start with maxlag as first estimate and increase iteratively. Defaults to TRUE. Of class logical. [-]}

\item{\code{nplag}}{"n" - consider negative lag times only, i.e. obs is expected to lag behind ref; "p" - consider positive lag times only, i.e. ref is expected to lag behind obs; "np" - consider negative and positive lag times. Defaults to "np". Of class character. [-]}

\item{\code{absolute}}{TRUE - consider positive and negative correlations when finding lag time; FALSE - consider positive correlations only when finding lag time. Defaults to TRUE. Of class logical. [-]}

\item{\code{hpf}}{TRUE - apply Butterworth high-pass filter; FALSE - use raw data. Defaults to TRUE. Of class logical. [-]}

\item{\code{FRACmin}}{Minimum fraction of data to attempt lag determination. Defaults to 0.1. Of class numeric. [-]}
}
\value{
Lagged input data and calculation results in a list consisting of:\cr
\code{refDATA} The reference data.
\code{obsDATA} The data that was lagged to coincide with the reference data.
\code{lag} The number of data points by which the lag correction was performed.
\code{ccf} The cross-correction coefficient between ref and obs for the determined lag time.
}
\description{
Function definition. Lag two datasets, so as to maximise their cross-correlation.
}
\examples{
Currently none.
}
\author{
Stefan Metzger \email{eddy4R.info@gmail.com}
}
\references{
License: GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007.
}
\seealso{
Currently none.
}
\keyword{aggregate,}
\keyword{average,}
\keyword{descriptive}
\keyword{statistics}

