% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def.neon.read.hdf5.para.R
\name{def.neon.read.hdf5.para}
\alias{def.neon.read.hdf5.para}
\title{Definition function: Function to read in dp0p files}
\usage{
def.neon.read.hdf5.para(DirFileParaLoca, GrpName, PosPara = NULL)
}
\arguments{
\item{DirFileParaLoca}{Absolute path to h5 file from which workflow and scientific parameters are to be used, of class "character". [-]}

\item{GrpName}{Absolute path to h5 file group and variable name from which parameters are to be read, of class "character". [-]}

\item{PosPara}{Parameters can be subset by PosPara; defaults to NULL which imports all parameters. A vector of characters. [-]}
}
\value{
The returned object is the parameters extracted from dp0p file for future use.
}
\description{
Definition function. A function to read in dp0p files and extact relevant attributes as parameters for future use.
}
\examples{
Currently none
}
\references{
License: GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007. \cr
}
\seealso{
Currently none.
}
\author{
Stefan Metzger \email{eddy4R.info@gmail.com} \cr
David Durden \email{ddurden@battelleecology.org} \cr
Hongyan Luo \email{hluo@battelleecology.org}
}
\keyword{dp0p}
\keyword{hdf5,}
\keyword{parameter,}
