% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def.rglr.R
\name{def.rglr}
\alias{def.rglr}
\title{Definition function: Regularizing irregular data to strictly regular / equidistant data}
\usage{
def.rglr(timeMeas, dataMeas, unitMeas = base::attributes(dataMeas)$unit,
  BgnRglr = NULL, EndRglr = NULL,
  TzRglr = base::attributes(BgnRglr)$tzone, FreqRglr,
  MethRglr = c("CybiEc", "cybiDflt", "zoo")[1], WndwRglr = c("Cntr",
  "Lead", "Trlg")[1], IdxWndw = c("Clst", "IdxWndwMin", "IdxWndwMax")[1],
  PrcsSec = 6)
}
\arguments{
\item{timeMeas}{A vector containing the observation times. Of class "POSIXlt" including timezone attribute, and of the same length as \code{dataMeas}. [-]}

\item{dataMeas}{A named data.frame containing the observations. Columns may be of class "numeric" or "integer", and of the same length as \code{timeMeas}. Columns of classes other than "numeric" or "integer" are removed and not included in the returned \code{dataRegl}. [user-defined]}

\item{unitMeas}{A vector containing the unit of each column in \code{dataMeas}. Of class "character". It is recommended to conform to the "unit representation" guidelines documented in the eddy4R.base package.}

\item{BgnRglr}{Desired begin time for the regularized dataset. Of class "POSIXlt" including timezone attribute, and \code{length(BgnRglr) = 1}. This input is not used in the "cybiDflt" method. [-]}

\item{EndRglr}{Desired end time for the regularized dataset. Of class "POSIXlt" including timezone attribute, and \code{length(EndRglr) = 1}. This input is not used in the "cybiDflt" method. [-]}

\item{TzRglr}{Desired timezone for the regularized dataset. Of class "character" and \code{length(TzRglr) = 1}, defaults to the same timezone as \code{BgnRglr}. For the "cybiDflt" method, the same time zone as timeMeas is used. [-]}

\item{FreqRglr}{Desired frequency of  the regularized dataset. Of class "numeric" or "integer" and \code{length(FreqRglr) = 1}. [Hz]}

\item{MethRglr}{Switch for different regularization methods. Of class "character", currently defaults to "CybiEc". [-] \cr
Method "cybiDflt" implements the default for metereological variable regularization performed by NEON CI. Namely, a new time series is created 
from the first measurement time, rounded toward zero, using the expected data frequency. The first measurement falling 
in between one time stamp and the next is assigned to the first of these, and all other measurements falling in this range are ignored.\cr
Method "CybiEc" implements the default regularization method for eddy-covariance processing utilized CI. The procedure 
is documented in NEON.DOC.001069.\cr
Method "zoo" implements the regularization method using the zoo::na.approx function. This method can only handle up to millisecond precision (PrcsSec=3)}

\item{WndwRglr}{Position of the window for binning in the "CybiEc" method. \code{WndwRglr} can be centered [Cntr], leading [Lead], or trailing [Trlg] (defaults to centered).\cr}

\item{IdxWndw}{Determines which observation to allocate to a bin if multiple observations fall into a single bin when using the "CybiEc" method.. \code{IdxWndw} can be set to closest [Clst], first [IdxWndwMin], or last [IdxWndwMax] (defaults to closest).\cr}

\item{PrcsSec}{A single numeric (integer) value indicating the operational precision of the seconds field of time vectors. Defaults to 6 (microsecond-precision). Values higher than 6 cannot be guaranteed to produce desired results.}
}
\value{
Returns a list with elements \code{TzRglr}, \code{FreqRglr}, \code{MethRglr}, \code{timeRglr}, and \code{dataRglr}.
}
\description{
Function defintion. 
Takes a (potentially) irregularly spaced timeseries \code{timeMeas} of data \code{dataMeas} and returns a strictuly regularly spaced timeseries \code{timeRglr} of data \code{dataRglr}. \strong{ATTENTION}: \code{MethRglr = "zoo"} uses the zoo:na.approx() function, which does not currently abide by its \code{maxgap} argument version 1.7-13. In result, where gaps exist currently the last known value is repeated instead of NAs being inserted. An Email with a request for bugfixing has been sent to \email{Achim.Zeileis@R-project.org} (2016-05-08).
}
\examples{
# make sure that fractional seconds can be seen from the console
options(digits.secs=3)
# assign measured time vector
timeMeas <- base::as.POSIXlt(seq.POSIXt(
  from = base::as.POSIXlt("2016-01-01 00:00:00.001", format="\%Y-\%m-\%d \%H:\%M:\%OS", tz="UTC"),
  to = base::as.POSIXlt("2016-01-01 00:00:01.002", format="\%Y-\%m-\%d \%H:\%M:\%OS", tz="UTC"),
  by = 1/10), tz = "UTC")[-c(5,6,8)]
# assign fake observations
dataMeas <- base::data.frame("wind01" = rnorm(base::length(timeMeas)), "wind02" = rnorm(base::length(timeMeas)))
# regularize
def.rglr(
  timeMeas = timeMeas,
  dataMeas = dataMeas,
  unitMeas = c("metersPerSecond", "metersPerSecond"),
  BgnRglr = base::as.POSIXlt("2016-01-01 00:00:00.000", format="\%Y-\%m-\%d \%H:\%M:\%OS", tz="UTC"),
  EndRglr = base::as.POSIXlt("2016-01-01 00:00:01.000", format="\%Y-\%m-\%d \%H:\%M:\%OS", tz="UTC"),
  FreqRglr = 10,
  MethRglr = "zoo"
)

#"CybiEc" example with multiple observations in a single bin
timeMeas[3] <- timeMeas[3] - 0.03
timeMeas[2] <- timeMeas[2] - 0.03
timeMeas[1] <- timeMeas[1] + 0.05
timeMeas[7] <- timeMeas[7] - 0.031
timeMeas[8] <- timeMeas[8] - 0.081

#Regularize with a centered window and chosing the closest value to the regularized timestamp. 
def.rglr(
  timeMeas = timeMeas,
  dataMeas = dataMeas,
  unitMeas = c("metersPerSecond", "metersPerSecond"),
  BgnRglr = base::as.POSIXlt("2016-01-01 00:00:00.000", format="\%Y-\%m-\%d \%H:\%M:\%OS", tz="UTC"),
  EndRglr = base::as.POSIXlt("2016-01-01 00:00:01.000", format="\%Y-\%m-\%d \%H:\%M:\%OS", tz="UTC"),
  FreqRglr = 10,
  MethRglr = "CybiEc",
  WndwRglr = "Cntr",
  IdxWndw = "Clst"
)  
}
\references{
License: GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007. \cr
NEON.DOC.001069 Preprocessing ATBD: The ATBD that describes the CybiEc and cybiDflt regularization methods. \cr
}
\seealso{
?zoo:na.approx, ?stats::approx
}
\author{
Stefan Metzger \email{eddy4R.info@gmail.com} \cr
Cove Sturtevant \email{eddy4R.info@gmail.com} \cr
David Durden \email{ddurden@battelleecology.org}
}
\keyword{equidistant,}
\keyword{preprocessing}
\keyword{regularization,}
