% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def.rot.enu.ned.R
\name{def.rot.enu.ned}
\alias{def.rot.enu.ned}
\title{Definition function: Coordinate transformation from the east-north-up (Enu) coordiante system to the north-east-down (Ned) coordiante system}
\usage{
def.rot.enu.ned(angEnu, Meth = c("vec", "ang"))
}
\arguments{
\item{angEnu}{Angle in east-north-up. [rad]}

\item{Meth}{A vector of class "character" containing the method to use vector or angular, Meth = c("vec", "ang"). [-]}
}
\value{
Azimuth angle in the North-east-down local coordinate system. [rad]
}
\description{
Function defintion. Coordinate transformation from the east-north-up (Enu) coordiante system to the north-east-down (Ned) coordiante system, and vice versa.
}
\examples{
Currently none.
}
\references{
License: GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007.
}
\seealso{
Currently none.
}
\author{
David Durden \email{ddurden@battelleecology.org}
}
\keyword{AMRS,}
\keyword{IMU,}
\keyword{coordinate,}
\keyword{transformation}
