% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def.ssi.mean.R
\name{def.ssi.mean}
\alias{def.ssi.mean}
\title{Calculation of average signal strength for LI-7200 IRGA}
\usage{
def.ssi.mean(ssiCo2, ssiH2o)
}
\arguments{
\item{\code{ssiCo2}}{A vector containing the CO2 signal strength, of class "numeric". [percent]}

\item{\code{ssiH2o}}{A vector containing the H2O signal strength, of class "numeric". [percent]}
}
\value{
The returned object is the average signal strength calculated from the CO2 signal strength and the H2O signal strength.
}
\description{
Function definition. Calculation of average signal strength for LI-7200 IRGA.
}
\examples{
Currently none.
}
\author{
Stefan Metzger \email{eddy4R.info@gmail.com}
Hongyan Luo \email{eddy4R.info@gmail.com}
}
\references{
Currently none.
}
\seealso{
Currently none.
}
\keyword{irga}
\keyword{signal}
\keyword{strength,}

