% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def.unit.extr.R
\name{def.unit.extr}
\alias{def.unit.extr}
\title{Definition function: Extract variable with unit attribute from data frame}
\usage{
def.unit.extr(data, nameVar, AllwPos = FALSE)
}
\arguments{
\item{\code{data}}{Required. A data frame with units of each variables attached as a named attribute "unit"
to the data frame.}

\item{\code{nameVar}}{Required. A string containing the name of the variable within \code{data} to extract,
along with its units}

\item{\code{AllwPos}}{Optional. Logical, defaulting to FALSE. Allow positional determination of the unit within
the unit attribute vector? If FALSE, an error will result if the variable name does not match an entry with 
the named unit vector. If TRUE, the unit attribute vector does not need to be named, and the unit will be 
pulled based on the same position of the variable within the data frame \code{data}.}
}
\value{
The desired variable with units attached as attribute.
}
\description{
Function definition. Extracts a variable from a data frame (with units attached as attributes
at the data frame level), and attaches the variable's unit directly to the extracted variable.
}
\examples{
Currently none
}
\author{
Cove Sturtevant \email{csturtevant@neoninc.org}
}
\references{
license: Terms of use of the NEON FIU algorithm repository dated 2015-01-16
}
\seealso{
Currently none
}
\keyword{Fulcrum}

