% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rot_B2M.r
\name{rot_B2M}
\alias{rot_B2M}
\title{Coordinate transformation from CSAT3 body coordinate system to meteorological coordiante system}
\usage{
rot_B2M(Psi_boom = eddy4R.base::def.conv.unit(data = 189.28, unitFrom = "deg",
  unitTo = "rad"), Psi_B2M_0 = eddy4R.base::def.conv.unit(data = 90, unitFrom
  = "deg", unitTo = "rad"), BCS_uvw = data.frame(u = 0, v = 0, w = 0))
}
\arguments{
\item{Currently}{none}
}
\value{
Currently none
}
\description{
Coordinate transformation from CSAT3 body coordinate system to meteorological coordiante system.
}
\examples{
Currently none
}
\author{
Stefan Metzger \email{eddy4R.info@gmail.com}
}
\references{
Currently none
}
\seealso{
Currently none
}
\keyword{Currently}
\keyword{none}

